### 2025-05-17 #8242 @arfon
Also, I noted your commentary at submission time about the paper length. I will defer to the track EiC here on whether they will allow this. Please be aware that the guideline for paper length is about reducing the amount of volunteer labour our reviewers and editors have to handle. The JOSS _software_ review is already an intensive process, and so while I am sympathetic to your point about not all of the details being available in open access publications, it also seems like you could address this issue with a pre-print publication somewhere.
