### 2025-05-23 #8242 @Kevin-Mattheus-Moerman
@ChrisWoodgate in line with @arfon's comments there, we will need to shorten the paper before we proceed. Please shorten to be closer to our guidelines. You can offload to a documentation document or preprint like document if you prefer.
