### 2025-05-23 #8242 @ChrisWoodgate
Hi @Kevin-Mattheus-Moerman, happy to work with you on this, but please could you be more specific with your definition of 'closer'. (I don't want to cut unnecessarily harshly, but also don't want to have to go through the process of shortening the paper more than once.)

I was a reviewer recently for a JOSS paper (which is now published) around 2000 words in length, and am aware of other accepted JOSS papers which are over 2500 words in length.

What length of paper would be acceptable to you?

Thanks!
