### 2025-05-23 #8242 @jgostick
Hi @ChrisWoodgate
I  will be the associate editor on this one. I took a look at the paper, and I can tell you that you should cut pretty much all the theory, so the following sections:

- Alloy Thermodynamics
- Evaluation of the alloy internal energy
- Sampling Algorithms (and related subsections).  This can be condensed to a single paragraph list the algorithm you offer along with some relevant citations. You don't need to explain the theory.
- Physical Quantities of Interest can be condensed to a single paragraph, again just listing what the package outputs without providing the theory.

Removal of these sections will also reduce the References by quite a bit, which are currently spanning nearly 4 pages, which is the length of some JOSS papers in their entirety.

The theory that you have is very nice, but should be moved to a page in your documentation.

Also:

- The Purpose of BraWL section should be moved into the Statement of Need section.
- The Example Applications sections seems ok since it is more of a practical demo of the software.
