### 2025-05-23 #8242 @ChrisWoodgate
Okay, thanks for the clarification @jgostick. I think we will probably go with the suggestion made by @Kevin-Mattheus-Moerman of offloading the theory to a preprint, as we already have one of these on the arXiv which we can point to in the article. This will also reduce the citations made in the article.

Hubert (first author) is away until the middle of next week, but once he is back I will discuss with him and the other authors and we will take it from there. You should hopefully hear back from us in the next week or two
