### 2025-05-28 8cdf91d
Modified to account for performance analysis
