### 2025-05-28 9b6fc15
Modified to account for performance analysis
