### 2025-05-28 f16991b
Bash script to submit slurm jobs
