### 2025-05-28 f43f7b0
Python script for generating slurm scripts
