### 2025-05-28 74818f1
Makes directories for performance analysis and populates them with appropriate input files
