+++ performance/make_directories.py
@@ -1,0 +1,49 @@
+import os
+import shutil
+
+# Function to create directories and copy only .inp files from the source directory
+def create_directories_and_copy_files(brawl_type, source_dir, num_walkers):
+    # Cores list for which directories will be created
+    windows = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16]
+
+    # Check if source directory exists
+    if not os.path.exists(source_dir):
+        print(f"Source directory '{source_dir}' does not exist.")
+        return
+
+    # Loop through the range of values for windows
+    for window in windows:
+      for i in range(3):
+          # Create the directory name following the naming convention
+          dir_name = f'{brawl_type}_{num_walkers:02d}_{window:02d}_{i+1}'
+
+          # Create the directory if it does not exist
+          if not os.path.exists(dir_name):
+              os.makedirs(dir_name)
+
+          for item in os.listdir(source_dir):
+              source_item = os.path.join(source_dir, item)
+              destination_item = os.path.join(dir_name, item)
+              
+              # Only copy .inp or .vij files, and skip directories
+              if os.path.isfile(source_item) and item.endswith(('.inp', '.vij')):
+                  shutil.copy2(source_item, destination_item)  # copy2 preserves metadata
+          
+          file_path = os.path.join(dir_name, "wl_input.inp")
+          with open(file_path, "r") as file:
+              lines = file.readlines()
+
+          for j in range(len(lines)):
+            if lines[j].strip().startswith("num_windows="):
+                lines[j] = f"num_windows={window}\n"
+            elif lines[j].strip().startswith("performance="):
+                lines[j] = f"performance={brawl_type}\n"
+
+          with open(file_path, "w") as file:
+              file.writelines(lines)
+
+# Example call to the function with brawl_type and source_dir
+brawl_type = 4
+source_dir = '.'
+num_walkers = 16
+create_directories_and_copy_files(brawl_type, source_dir, num_walkers)
