### 2025-06-03 8717a10
Implemented new energy biasing for moving configurations into energy windows
