### 2025-06-03 3b84dba
Removed unused variable
