### 2025-06-04 56ddaa0
Reducing level of detail in paper in line with editorial guidance
