+++ paper/paper.md
@@ -56,65 +56,19 @@ The Fortran package `BraWl` facilitates simulation of the thermodynamics and pha
 It achieves this by providing implementation of both the Bragg-Williams Hamiltonian (a lattice based model expressing the internal energy of an alloy as a sum of atom-atom effective pair interactions) concurrently with a range of conventional and enhanced sampling techniques for exploration of the alloy configuration space.
 The result is a package which can determine phase equilibria as a function of both temperature and alloy composition, which leads to the construction of alloy phase diagrams.
 Additionally, the package can be used for extraction of representative equilibrated atomic configurations for visualisation, as well as for use in complementary modelling approaches.
-Before outlining the specific need for this package, we first review some elementary details of alloy thermodynamics and statistical mechanics.
+An in-depth discussion of the background, underlying theory, and technical details of the sampling algorithms implemented in the package is given in @naguszewskibrawl2025.
 
-## Alloy thermodynamics
+There are a range of existing packages capable of simulating alloy phase equilibria, both open- and closed-source.
+Examples of widely-used such packages include ATAT [@vandewallealloy2002], ICET [@angqvisticet2019] and CELL [@rigamonticell2024], though all of these focus primarily on implementation of a generalised cluster expansion, rather than the simpler form of the Bragg-Williams Hamiltonian.
+To our knowledge, there is no open-source package specifically focussing on the implementation of a range of sampling algorithms applied to the Bragg-Williams model.
+We therefore believe that BraWl fills a gap in the capabilities of the current alloy software ecosystem.
+Additionally, we hope that the modular way in which the package is constructed could enable implementation of more complex Hamiltonians, as well as further sampling algorithms in addition to those detailed below, in due course.
 
-In a substitutional alloy with fixed underlying lattice, a particular arrangement of atoms can be specified by a discrete set of site occupation numbers, $\left\{ \xi_{i\gamma} \right\}$, where $\xi_{i \gamma} = 1$ if site $i$ is occupied by an atom of species $\gamma$ and $\xi_{i \gamma} = 0$ otherwise.
-The lattice index $i$ takes values in the range 1 to $N$, where $N$ is the total number of lattice sites in the system, while the species index $\gamma$ takes values in the range 1 to $s$, where $s$ is the number of chemical species (elements) present in the alloy composition.
-The physical constraint that each lattice site is occupied by one (and only one)  atom is expressed as $\sum_{\gamma} \xi_{i\gamma} = 1$, while the total concentration of a given chemical species is given by $c_\gamma = \frac{1}{N} \sum_i \xi_{i\gamma}$.
-(Naturally, vacancies can be treated in this framework by considering them as an additional chemical species present at a very low concentration.)
-If we consider an ensemble of alloy configurations, we can then define the site-wise concentrations, $\left\{ c_{i\gamma}\right\}$, as the average value of the site occupation numbers across the ensemble, $c_{i\gamma} = \langle \xi_{i\gamma} \rangle$, where $\langle \cdot \rangle$ denotes the average taken over the ensemble.
+## The Bragg-Williams Hamiltonian
 
-In the context of substitutional alloys, the word _phase_ refers to a physically homogeneous area of the material with uniform chemical composition and physical properties.
-At the atomic level, this means that, on the average, the site-wise concentrations in a particular spatial region are homogeneous and describe some repeating motif of atoms arranged on a lattice.
[…]
