+++ paper/brawl.bib
@@ -1,4 +1,3 @@
-
 @article{marchantab2021,
 	title = {\textit{{Ab} initio} calculations of the phase behavior and subsequent magnetostriction of {Fe}$_{\textrm{1−x}}${Ga}$_{\textrm{x}}$ within the disordered local moment picture},
 	volume = {103},
@@ -54,7 +53,7 @@
 	number = {3},
 	urldate = {2024-03-23},
 	journal = {Physical Review Materials},
-	author = {Shenoy, Lakshmi and Woodgate, Christopher D. and Staunton, Julie B. and Bartok, Albert P. and Becquart, Charlotte S. and Domain, Christophe and Kermode, James R.},
+	author = {Shenoy, Lakshmi and Woodgate, Christopher D. and Staunton, Julie B. and Bartók, Albert P. and Becquart, Charlotte S. and Domain, Christophe and Kermode, James R.},
 	month = mar,
 	year = {2024},
 	pages = {033804},
@@ -146,56 +145,12 @@
 	pages = {2881--2898},
 }
 
-@book{woodgatemodelling2024,
-	address = {Cham},
-	series = {Springer {Series} in {Materials} {Science}},
-	title = {Modelling {Atomic} {Arrangements} in {Multicomponent} {Alloys}: {A} {Perturbative}, {First}-{Principles}-{Based} {Approach}},
-	volume = {346},
-	copyright = {https://www.springernature.com/gp/researchers/text-and-data-mining},
-	isbn = {978-3-031-62020-1 978-3-031-62021-8},
-	shorttitle = {Modelling {Atomic} {Arrangements} in {Multicomponent} {Alloys}},
-	url = {https://doi.org/10.1007/978-3-031-62021-8},
-	urldate = {2024-08-12},
-	publisher = {Springer Nature Switzerland},
-	author = {Woodgate, Christopher D.},
-	year = {2024},
-}
-
-@article{ferrarisimulating2023,
-	title = {Simulating short-range order in compositionally complex materials},
-	volume = {3},
-	issn = {2662-8457},
-	url = {https://doi.org/10.1038/s43588-023-00407-4},
-	doi = {10.1038/s43588-023-00407-4},
-	number = {3},
-	urldate = {2024-04-14},
-	journal = {Nature Computational Science},
-	author = {Ferrari, Alberto and Körmann, Fritz and Asta, Mark and Neugebauer, Jörg},
-	month = mar,
-	year = {2023},
-	pages = {221--229},
-}
-
-@article{ferrarifrontiers2020,
-	title = {Frontiers in atomistic simulations of high entropy alloys},
-	volume = {128},
-	issn = {0021-8979, 1089-7550},
-	url = {https://doi.org/10.1063/5.0025310},
-	doi = {10.1063/5.0025310},
-	number = {15},
-	urldate = {2022-01-06},
-	journal = {Journal of Applied Physics},
-	author = {Ferrari, Alberto and Dutta, Biswanath and Gubaev, Konstantin and Ikeda, Yuji and Srinivasan, Prashanth and Grabowski, Blazej and Körmann, Fritz},
-	month = oct,
-	year = {2020},
-	pages = {150901},
-}
-
 @book{landauguide2014,
 	address = {Cambridge, UK},
 	edition = {4},
 	title = {A {Guide} to {Monte} {Carlo} {Simulations} in {Statistical} {Physics}},
 	url = {https://doi.org/10.1017/CBO9781139696463},
+    doi = {10.1017/CBO9781139696463},
 	publisher = {Cambridge University Press},
 	author = {Landau, David P. and Binder, Kurt},
 	year = {2014},
@@ -261,21 +216,6 @@
[…]
