### 2025-06-04 511e7be
Fixing some minor typsetting errors
