+++ paper/paper.md
@@ -56,7 +56,7 @@ The Fortran package `BraWl` facilitates simulation of the thermodynamics and pha
 It achieves this by providing implementation of both the Bragg-Williams Hamiltonian (a lattice based model expressing the internal energy of an alloy as a sum of atom-atom effective pair interactions) concurrently with a range of conventional and enhanced sampling techniques for exploration of the alloy configuration space.
 The result is a package which can determine phase equilibria as a function of both temperature and alloy composition, which leads to the construction of alloy phase diagrams.
 Additionally, the package can be used for extraction of representative equilibrated atomic configurations for visualisation, as well as for use in complementary modelling approaches.
-An in-depth discussion of the background, underlying theory, and technical details of the sampling algorithms implemented in the package is given in @naguszewskibrawl2025.
+An in-depth discussion of the background, underlying theory, and technical details of the sampling algorithms implemented in the package is given by @naguszewskibrawl2025.
 
 There are a range of existing packages capable of simulating alloy phase equilibria, both open- and closed-source.
 Examples of widely-used such packages include ATAT [@vandewallealloy2002], ICET [@angqvisticet2019] and CELL [@rigamonticell2024], though all of these focus primarily on implementation of a generalised cluster expansion, rather than the simpler form of the Bragg-Williams Hamiltonian.
@@ -98,14 +98,16 @@ This is the purpose of `BraWl` as presented in this work.
 `BraWl` implements a range of conventional and enhanced sampling algorithms for exploration of the alloy configuration space.
 Throughout, the package defaults to performing _swaps_ of pairs of atoms in the simulation cell, to conserve the overall concentration of each chemical species present in the simulation.
 At present, the implemented sampling algorithms are:
+
 1. **The Metropolis-Hastings Monte Carlo algorithm.** The Metropolis-Hastings algorithm allows for a system of interest to follow a chain of states which evolve to, and sample, an equilibrium ensemble [@metropolisequation1953; landauguide2014]. It is a useful method for obtaining the equilibrium state of a system at a given temperature.
 2. **Wang-Landau sampling.** Wang-Landau sampling is a flat histogram method which provides a means for high throughput calculation of phase diagrams for atomistic/lattice model systems [@wangefficient2001]. The method allows for direct computation of an estimate of the density of states in energy, $g(E)$, and hence the partition function of a system of interest. From the partition function, thermodynamic quantities at any temperature of interest can then be obtained provided one has prior knowledge of the minimum and maximum energy relevant to those temperatures.
-3. **Nested Sampling.** Nested sampling (NS) is powerful Bayesian inference technique [@skillingnested2004; @skillingnested2006] adapted to sample the potential energy surface of atomistic systems [@ashtonnested2022], giving direct access to the partition function at arbitrary temperatures for comprehensive thermodynamic analysis, without relying on advance knowledge of relevant structures or the range of energies accessible to them [@partayefficient2010; partaynested2021].
[…]
