+++ paper/brawl.bib
@@ -232,7 +232,7 @@
 	pages = {271},
 }
 
-@article{woodgateintegrated2024,
+@article{woodgateint2024,
 	title = {Integrated \textit{ab initio} modelling of atomic ordering and magnetic anisotropy for design of {FeNi}-based magnets},
 	volume = {10},
 	issn = {2057-3960},
