### 2025-06-04 aceb297
Fixing some additional typsetting issues and moving figure
