+++ paper/brawl.bib
@@ -315,7 +315,7 @@
 	number = {32},
 	urldate = {2025-01-21},
 	journal = {The Journal of Physical Chemistry B},
-	author = {Pártay, Lívia B. and Bartók, Albert P. and Csányi, Gábor},
+	author = {Pártay, Livia B. and Bartók, Albert P. and Csányi, Gábor},
 	month = aug,
 	year = {2010},
 	pages = {10502--10512},
