### 2025-06-04 #8346 @editorialbot
[REVIEW]: BraWl: Simulating the thermodynamics and phase stability of multicomponent alloys using conventional and enhanced sampling techniques
**Submitting author:** @ChrisWoodgate (<a href="http://orcid.org/0000-0003-4876-5558">Christopher D. Woodgate</a>)
**Repository:** https://github.com/ChrisWoodgate/BraWl
**Branch with paper.md** (empty if default branch): joss_review
**Version:** v1.0.1
