### 2025-06-04 f833dad
Fixed initial array overlap creation
