### 2025-06-12 #8346 @ChrisWoodgate
Hi @amkrajewski, @sudarshanv01,

Thanks very much for agreeing to review our submission. In case you are not familiar with the JOSS review process, it starts by asking `editorialbot` to generate a reviewer checklist using the command
which generates a list of things for you to check about both the paper and software repository.

If you have any issues with compilation of the code or running the examples, please just drop me or @HubertJN (first author) a message.

Cheers! 🙂
