### 2025-06-12 a571133
Fixing typo in README.md
