### 2025-06-27 ChrisWoodgate/BraWl#49 @sudarshanv01
Data for figure in JOSS manuscript?
Hello, reviewing the paper for JOSS [here](https://github.com/openjournals/joss-reviews/issues/8346).

One of the points of the review is:

> Data sharing: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.

I see you have applied the code in Figures 1 and 3, do you have the data for those figures stored somewhere public?
