### 2025-06-27 ChrisWoodgate/BraWl#49 @ChrisWoodgate
Thanks for pointing this out. The scripts for generating these figures are provided in the `examples` subdirectory of the package. We _could_ provide data as well, but I am of the view that scripts to reproduce the plots are sufficient. What are your thoughts on this?

P.S. I am about to head out of the office so apologies if I don't respond immediately to further comments.
