### 2025-07-01 77f87e4
added option for Kawasaki swaps
