### 2025-07-01 c769e4c
Added functionality to add slurm job id to random seed if available
