+++ src/wang-landau.F90
@@ -103,6 +103,13 @@ module wang_landau
     setup_internal = setup
     wl_setup_internal = wl_setup
 
+    ! Selects MC atom selection critetion
+    if (wl_setup_internal%nbr_swap) then
+      wl_setup_internal%mc_select => select_sites_nbr
+    else
+      wl_setup_internal%mc_select => select_sites
+    end if
+
     ! Check if number of MPI processes is divisible by number of windows
     call MPI_COMM_SIZE(MPI_COMM_WORLD, mpi_processes, ierr)
     print*, "MPI processes: ", mpi_processes
@@ -511,8 +518,7 @@ module wang_landau
     do i = 1, wl_setup_internal%mc_sweeps*setup_internal%n_atoms
       ! Make one MC trial
       ! Generate random numbers
-      rdm1 = setup_internal%rdm_site()
-      rdm2 = setup_internal%rdm_site()
+      call wl_setup_internal%mc_select(rdm1, rdm2)
       ! Get what is on those sites
       site1 = config(rdm1(1), rdm1(2), rdm1(3), rdm1(4))
       site2 = config(rdm2(1), rdm2(2), rdm2(3), rdm2(4))
@@ -707,6 +713,7 @@ module wang_landau
   !> @date    2024 
   subroutine pre_sampling(wl_logdos, mpi_wl_hist)
     real(real64), allocatable, intent(inout) :: wl_logdos(:), mpi_wl_hist(:)
+    i_sweeps = 0
     pre_sampled = 0
     pre_sampled_buffer = 0
     pre_sampled_state = 0
@@ -715,6 +722,11 @@ module wang_landau
     end = start
     do while (SUM(pre_sampled_buffer) < wl_setup_internal%num_windows)
       call sweeps(wl_logdos, mpi_wl_hist)
+      i_sweeps = i_sweeps + 1
+      if (MOD(i_sweeps, 10) == 0) then
+        i_sweeps = 1
+        call replica_exchange(config)
+      end if
 
       if (minval(mpi_wl_hist) > REAL(setup_internal%n_atoms) .and. pre_sampled_state == 0) then
         pre_sampled(mpi_index) = 1
@@ -835,18 +847,18 @@ module wang_landau
     integer :: i
 
     ! Bins
-    integer :: bins(wl_setup_internal%num_windows)
+    integer :: bins
 
-    bins = window_intervals(:,2)-window_intervals(:,1)
-
-    window_indices(1, 1) = window_intervals(1,1)
-    window_indices(1,2) = INT(window_intervals(1,2) + wl_setup_internal%bin_overlap*bins(2))
+    window_indices(1,1) = window_intervals(1,1)
+    window_indices(1,2) = window_intervals(1,2)
     do i = 2, wl_setup_internal%num_windows-1
-      window_indices(i, 1) = MAX(INT(window_intervals(i,1) - wl_setup_internal%bin_overlap*bins(i-1)), 1)
-      window_indices(i, 2) = MIN(INT(window_intervals(i,2) + wl_setup_internal%bin_overlap*bins(i+1)), wl_setup_internal%bins)
+      bins = window_indices(i-1,2) - window_indices(i-1, 1)
+      window_indices(i, 1) = MAX(INT(window_intervals(i,1) - CEILING(wl_setup_internal%bin_overlap*bins)), 1)
+      window_indices(i, 2) = window_intervals(i,2)
     end do
+    bins = window_indices(wl_setup_internal%num_windows-1,2) - window_indices(wl_setup_internal%num_windows-1, 1)
     window_indices(wl_setup_internal%num_windows, 1) = MAX(INT(window_intervals(wl_setup_internal%num_windows,1) &
[…]
