### 2025-07-09 c446d02
Updates to README following JOSS feedback
