+++ src/wang-landau.F90
@@ -59,7 +59,8 @@ module wang_landau
   logical :: rho_saved
   
   ! Load balancing metrics
-  real(real64), allocatable :: lb_bins(:,:), lb_avg_time(:,:), lb_max_time(:,:), window_time(:), diffusion_prev(:)
+  real(real64), allocatable :: lb_bins(:,:), lb_avg_time(:,:), lb_max_time(:,:), lb_mc_steps(:), lb_mc_steps_buffer(:)
+  real(real64), allocatable :: window_time(:), diffusion_prev(:)
   integer :: converged, converged_sum
   
   ! Radial density across energy
@@ -112,7 +113,9 @@ module wang_landau
 
     ! Check if number of MPI processes is divisible by number of windows
     call MPI_COMM_SIZE(MPI_COMM_WORLD, mpi_processes, ierr)
-    print*, "MPI processes: ", mpi_processes
+    if (my_rank == 0) then
+      print*, "MPI processes: ", mpi_processes
+    end if
     bins = wl_setup_internal%bins
     if (MOD(mpi_processes, wl_setup_internal%num_windows) /= 0) then
       if (my_rank == 0) then
@@ -149,9 +152,9 @@ module wang_landau
     !---------!
     ! Burn in !
     !---------!
-    
+    iter = -1
     call enter_energy_window()
-    print*, "Rank: ", my_rank, "within energy window"
+    iter = 0
   
     call comms_wait()
     flush(6)
@@ -185,6 +188,7 @@ module wang_landau
     mpi_wl_hist = 0.0_real64
     i = 0
     i_sweeps = 0
+    lb_mc_steps = 0.0_real64
     start = mpi_wtime()
     do while (wl_f > wl_setup_internal%tolerance)
       i_sweeps = i_sweeps + 1
@@ -201,6 +205,7 @@ module wang_landau
         end if
       end if 
       call sweeps(wl_logdos, mpi_wl_hist)
+      lb_mc_steps(iter+1) = lb_mc_steps(iter+1) + wl_setup_internal%mc_sweeps*setup_internal%n_atoms
       if (MOD(i_sweeps, 10) == 0) then
         call replica_exchange(config)
       end if
@@ -254,7 +259,7 @@ module wang_landau
         call save_rho_E(rho_saved, radial_record)
 
         call save_wl_data(bin_edges, wl_logdos, wl_hist)
-        call save_load_balance_data(window_indices, rank_time_buffer)
+        call save_load_balance_data(window_indices, rank_time_buffer, lb_mc_steps)
         
         if (ANY([0,1] == wl_setup_internal%performance)) then
         call mpi_window_optimise(iter)
@@ -399,9 +404,10 @@ module wang_landau
   !>
   !> @author  H. J. Naguszewski
   !> @date    2024 
-  subroutine save_load_balance_data(window_indices, rank_time_buffer)
+  subroutine save_load_balance_data(window_indices, rank_time_buffer, lb_mc_steps)
     integer, intent(in) :: window_indices(:, :)
-    real(real64), intent(in) :: rank_time_buffer(:, :)
+    real(real64), intent(in) :: rank_time_buffer(:, :), lb_mc_steps(:)
+    call MPI_REDUCE(lb_mc_steps, lb_mc_steps_buffer, num_iter, MPI_DOUBLE_PRECISION, MPI_SUM, 0, MPI_COMM_WORLD, ierr)
     if (my_rank == 0) then
       lb_bins(iter, :) = REAL(window_indices(:, 2) - window_indices(:, 1) + 1)
       lb_avg_time(iter, :) = rank_time_buffer(:,1)
@@ -411,6 +417,7 @@ module wang_landau
[…]
