### 2025-07-14 c61ee9a
Adjusted parameters
