### 2025-07-14 5fede99
Created for comparing core efficiency across parallelism methods
