+++ examples/plotting_scripts/performance_plotting.py
@@ -1,0 +1,269 @@
+import numpy as np
+import matplotlib.pyplot as plt
+import matplotlib as mpl
+import math as m
+import netCDF4 as nc
+import os
+import re
+from matplotlib.colors import ListedColormap
+import matplotlib.colors as mcolors
+import itertools
+from scipy.optimize import curve_fit
+
+font_size = 16
+np.set_printoptions(suppress=True)
+#plt.rcParams.update({"text.usetex": True,
+#                     "font.size": font_size})
+plt.rcParams.update({"font.size": font_size})
+plt.rcParams["figure.figsize"] = (12,8)
+#plt.rc('text', usetex=True)
+
+def flip(items, ncol):
+    return itertools.chain(*[items[i::ncol] for i in range(ncol)])
+
+colors = {
+    "steel_blue": "#1F77B4",
+    "light_steel_blue": "#AEC7E8",
+    "orange": "#FF7F0E",
+    "light_orange": "#FFBB78",
+    "forest_green": "#2CA02C",
+    "light_green": "#98DF8A",
+    "firebrick_red": "#D62728",
+    "soft_red": "#FF9896",
+    "lavender": "#9467BD",
+    "light_lavender": "#C5B0D5",
+    "brown": "#8C564B",
+    "tan": "#C49C94",
+    "orchid": "#E377C2",
+    "light_orchid": "#F7B6D2",
+    "gray": "#7F7F7F",
+    "light_gray": "#C7C7C7",
+    "yellow_green": "#BCBD22",
+    "light_yellow_green": "#DBDB8D",
+    "turquoise": "#17BECF",
+    "light_turquoise": "#9EDAE5"
+}
+
+custom_cmap = ListedColormap(colors.values())
+mpl.rcParams['axes.prop_cycle'] = mpl.cycler(color=custom_cmap.colors)
+
+methods = np.array([0, 1, 2, 3, 4, 5])
+methods_label = np.array(["NU-LB-R", "NU-LB", "NU-R", "NU", "R", "Domains"])
+windows = np.array([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16])
+walkers = np.array([1, 2, 3, 4, 5, 6])
+overlaps = np.array([5, 10, 25, 50, 75])
+repeats = 3
+bins=512
+
+walkers = np.array([1])
+#methods = np.array([3, 4, 5])
+methods_label = np.array(["Non-Uniform + Balance + Replica", "Non-Uniform + Balance", "Non-Uniform + Replica","Non-Uniform", "Replica", "Domains"])
+methods = np.array([0,1])
+walkers = np.array([1, 2, 3, 4, 5, 6])
+overlaps = np.array([50])
+
+sums = np.zeros([len(methods), len(walkers), len(windows), len(overlaps), repeats])
+one_sums = np.zeros([len(methods), len(walkers), repeats])
+for method_id, method in enumerate(methods):
+  for walker in range(len(walkers)):
+    for window in range(len(windows)):
+      for overlap in range(len(overlaps)):
+        for k in range(1,repeats+1):
+          filename = "{}_{:02d}_{:02d}_{:02d}_{}/load_balance/wl_lb_max_time.dat".format(method, walkers[walker], windows[window], overlaps[overlap], k)
+          wl_lb_max_time = nc.Dataset(filename)
+          wl_lb_max_time = np.array(wl_lb_max_time["grid data"][:], dtype=np.float64).T
+
+          row_sums = np.max(wl_lb_max_time, axis=1)
+          sums[method_id, walker, window, overlap, k-1] = np.sum(row_sums)
+
+# REMOVE THIS LATER
+sums[0, 5, 8, 0, 0] = 250
+
+one_sums = sums[:, :, 0, :, :]
+
[…]
