### 2025-07-14 5e50a34
Removed previous comment. Added extra LDFLAGs to adress compilation issue on other machines
