### 2025-08-04 9f4efd9
(tests) added serial tests on MacOS to the Actions CI
