+++ .github/workflows/testingOnPush_Apple.yaml
@@ -1,0 +1,43 @@
+name: MacOS (M1)
+
+on:
+  push:
+    paths:
+      - 'tests/**'
+      - 'src/**'
+      - '.github/workflows/testingOnPush_Apple.yaml'
+  workflow_dispatch:
+
+jobs:
+  test:
+    runs-on: macos-14
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Install gfortran and other dependencies with Homebrew
+        run: brew install gcc make pkg-config openblas netcdf netcdf-fortran open-mpi
+
+      - name: Install Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.11'
+
+      - name: Install Python dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r requirements.txt
+
+      - name: Run serial compilation with gfortran
+        run: |
+          make compiler=gfortran
+
+      - name: Run tests with gfortran
+        run: |
+          make compiler=gfortran tests     
+          cd tests
+          ../tests.run
+    
+      #- name: Run parallel compilation with mpifort
+      #  run: |
+      #    make compiler=mpifort
