### 2025-08-04 bc46736
(tests) minor CI imporvements
