+++ .github/workflows/testingOnPush_Apple.yaml
@@ -5,6 +5,8 @@ on:
     paths:
       - 'tests/**'
       - 'src/**'
+      - 'Makefile'
+      - 'requirements.txt'
       - '.github/workflows/testingOnPush_Apple.yaml'
   workflow_dispatch:
 
@@ -16,7 +18,7 @@ jobs:
         uses: actions/checkout@v4
 
       - name: Install gfortran and other dependencies with Homebrew
-        run: brew install gcc make pkg-config openblas netcdf netcdf-fortran open-mpi
+        run: brew install gcc make openblas netcdf netcdf-fortran open-mpi
 
       - name: Install Python
         uses: actions/setup-python@v5
