### 2025-08-04 869658d
(tests) added serial tests on Linux (Ubuntu x86_64) to the Actions CI
