+++ .github/workflows/testingOnPush_Ubuntu.yaml
@@ -1,0 +1,45 @@
+name: Ubuntu (x86_64)
+
+on:
+  push:
+    paths:
+      - 'tests/**'
+      - 'src/**'
+      - 'Makefile'
+      - 'requirements.txt'
+      - '.github/workflows/testingOnPush_Ubuntu.yaml'
+  workflow_dispatch:
+
+jobs:
+  test:
+    runs-on: ubuntu-22.04
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Install gfortran and other dependencies with apt
+        run: sudo apt-get install -y gfortran make libopenblas-dev libnetcdf-dev libnetcdf-fortran-dev mpich
+
+      - name: Install Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.11'
+
+      - name: Install Python dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r requirements.txt
+
+      - name: Run serial compilation with gfortran
+        run: |
+          make compiler=gfortran
+
+      - name: Run tests with gfortran
+        run: |
+          make compiler=gfortran tests     
+          cd tests
+          ../tests.run
+    
+      #- name: Run parallel compilation with mpifort
+      #  run: |
+      #    make compiler=mpifort
