### 2025-08-04 fdc4291
Update generate_documentation.yml
