+++ .github/workflows/generate_documentation.yml
@@ -7,6 +7,6 @@ jobs:
       - uses: mattnotmitt/doxygen-action@1.12.0
         with:
           working-directory: './'
-          doxyfile-path: 'docs/'
+          doxyfile-path: 'docs/doxyfile'
           enable-latex: true
       - run: doxygen docs/doxyfile
