### 2025-08-04 419dbcc
Update generate_documentation.yml
