+++ .github/workflows/generate_documentation.yml
@@ -6,7 +6,6 @@ jobs:
     steps:
       - uses: mattnotmitt/doxygen-action@1.12.0
         with:
-          working-directory: './'
-          doxyfile-path: 'docs/doxyfile'
+          doxyfile-path: './docs/doxyfile'
           enable-latex: true
       - run: doxygen docs/doxyfile
