+++ .github/workflows/generate_documentation.yml
@@ -6,6 +6,6 @@ jobs:
     steps:
       - uses: mattnotmitt/doxygen-action@1.12.0
         with:
-          doxyfile-path: './docs/doxyfile'
+          doxyfile-path: 'docs/doxyfile'
           enable-latex: true
       - run: doxygen docs/doxyfile
