### 2025-08-04 c6e78c2
Update generate_documentation.yml
