+++ .github/workflows/generate_documentation.yml
@@ -4,8 +4,7 @@ jobs:
   doxygen:
     runs-on: ubuntu-latest
     steps:
-      - uses: mattnotmitt/doxygen-action@1.12.0
-        with:
-          doxyfile-path: 'docs/doxyfile'
-          enable-latex: true
-      - run: doxygen docs/doxyfile
+      - name: "Doxygen"
+        run: |
+          sudo apt-get install -y doxygen
+          doxygen docs/doxyfile
