### 2025-08-04 11e6f92
Update generate_documentation.yml
