+++ .github/workflows/generate_documentation.yml
@@ -6,5 +6,6 @@ jobs:
     steps:
       - name: "Doxygen"
         run: |
+          pwd
           sudo apt-get install -y doxygen
           doxygen docs/doxyfile
