### 2025-08-04 f465c2b
Update generate_documentation.yml
