+++ .github/workflows/generate_documentation.yml
@@ -7,5 +7,6 @@ jobs:
       - name: "Doxygen"
         run: |
           pwd
+          ls
           sudo apt-get install -y doxygen
           doxygen docs/doxyfile
