### 2025-08-04 4095405
Update generate_documentation.yml
