### 2025-08-04 d0939f0
Update generate_documentation.yml
