+++ .github/workflows/generate_documentation.yml
@@ -6,6 +6,6 @@ jobs:
     steps:
       - name: "Doxygen"
         run: |
-          ls
+          ls $GITHUB_WORKSPACE
           sudo apt-get install -y doxygen
           doxygen docs/doxyfile
