### 2025-08-04 d1a7f2b
Update generate_documentation.yml
