### 2025-08-04 3fb54cf
Update generate_documentation.yml
