### 2025-08-04 94a5dcc
Update generate_documentation.yml
