+++ .github/workflows/generate_documentation.yml
@@ -4,9 +4,8 @@ jobs:
   doxygen:
     runs-on: ubuntu-latest
     steps:
-      - run: ls
-      - run: pwd
-      - name: "Doxygen"
-        run: |
-          sudo apt-get install -y doxygen
-          doxygen docs/doxyfile
+      - uses: actions/checkout@v4
+      - name: generate-documentation
+        uses: mattnotmitt/doxygen-action@1.12.0
+        with:
+          doxyfile-path: 'docs/doxyfile'
