### 2025-08-04 0d02719
(tests) small fix for dependency install on Linux
