+++ .github/workflows/testingOnPush_Ubuntu.yaml
@@ -18,7 +18,7 @@ jobs:
         uses: actions/checkout@v4
 
       - name: Install gfortran and other dependencies with apt
-        run: sudo apt-get install -y gfortran make libopenblas-dev libnetcdf-dev libnetcdf-fortran-dev mpich
+        run: sudo apt-get install -y gfortran make libopenblas-dev libnetcdf-dev libnetcdff-dev mpich
 
       - name: Install Python
         uses: actions/setup-python@v5
