### 2025-08-04 1051a5c
(tests) added *parallel* tests on Linux (Ubuntu x86_64) to the Actions CI
