### 2025-08-05 e08b36a
Added overlap tracker and modified overlap to function properly with minimum 2 bin overlap
