+++ src/wang-landau.F90
@@ -35,7 +35,7 @@ module wang_landau
   integer :: mpi_bins, mpi_start_idx, mpi_end_idx, mpi_index, mpi_start_idx_buffer
   integer :: mpi_end_idx_buffer, i_sweeps
   real(real64) :: scale_factor, scale_count, wl_logdos_min, bin_overlap, beta_diff, beta_original, beta_merge
-  real(real64), allocatable :: mpi_bin_edges(:), mpi_wl_hist(:), wl_logdos_buffer(:), wl_logdos_combine(:)
+  real(real64), allocatable :: mpi_bin_edges(:), mpi_wl_hist(:), wl_logdos_buffer(:), wl_logdos_combine(:), window_overlap(:,:)
   real(real64), allocatable :: rank_time(:), rank_time_buffer(:,:)
  
   ! Window variables
@@ -259,7 +259,7 @@ module wang_landau
         call save_rho_E(rho_saved, radial_record)
 
         call save_wl_data(bin_edges, wl_logdos, wl_hist)
-        call save_load_balance_data(window_indices, rank_time_buffer, lb_mc_steps)
+        call save_load_balance_data(window_indices, rank_time_buffer, lb_mc_steps, window_overlap)
         
         if (ANY([0,1] == wl_setup_internal%performance)) then
         call mpi_window_optimise(iter)
@@ -404,9 +404,9 @@ module wang_landau
   !>
   !> @author  H. J. Naguszewski
   !> @date    2024 
-  subroutine save_load_balance_data(window_indices, rank_time_buffer, lb_mc_steps)
+  subroutine save_load_balance_data(window_indices, rank_time_buffer, lb_mc_steps, window_overlap)
     integer, intent(in) :: window_indices(:, :)
-    real(real64), intent(in) :: rank_time_buffer(:, :), lb_mc_steps(:)
+    real(real64), allocatable, intent(in) :: rank_time_buffer(:, :), lb_mc_steps(:), window_overlap(:,:)
     call MPI_REDUCE(lb_mc_steps, lb_mc_steps_buffer, num_iter, MPI_DOUBLE_PRECISION, MPI_SUM, 0, MPI_COMM_WORLD, ierr)
     if (my_rank == 0) then
       lb_bins(iter, :) = REAL(window_indices(:, 2) - window_indices(:, 1) + 1)
@@ -416,6 +416,7 @@ module wang_landau
       call ncdf_writer_2d("load_balance/wl_lb_bins.dat", ierr, lb_bins)
       call ncdf_writer_2d("load_balance/wl_lb_avg_time.dat", ierr, lb_avg_time)
       call ncdf_writer_2d("load_balance/wl_lb_max_time.dat", ierr, lb_max_time)
+      call ncdf_writer_2d("load_balance/wl_window_overlap.dat", ierr, window_overlap)
       call ncdf_writer_1d("load_balance/wl_window_time.dat", ierr, window_time)
       call ncdf_writer_1d("load_balance/wl_lb_mc_steps.dat", ierr, lb_mc_steps_buffer)
     end if
@@ -850,22 +851,8 @@ module wang_landau
   !> @date    2024 
   subroutine create_window_intervals(window_intervals, window_indices, mpi_bins)
     integer, intent(inout) :: window_intervals(:,:), window_indices(:,:), mpi_bins
-    integer :: i
 
-    ! Bins
-    integer :: bins
-
-    window_indices(1,1) = window_intervals(1,1)
-    window_indices(1,2) = window_intervals(1,2)
-    do i = 2, wl_setup_internal%num_windows-1
-      bins = window_indices(i-1,2) - window_indices(i-1, 1)
-      window_indices(i, 1) = MAX(INT(window_intervals(i,1) - CEILING(wl_setup_internal%bin_overlap*bins)), 1)
[…]
