### 2025-08-05 393e2fd
Updated for new overlaps
