+++ performance/wl_vis.py
@@ -1,317 +1,55 @@
-import numpy as np
-import matplotlib.pyplot as plt
-import matplotlib as mpl
-import math as m
-import netCDF4 as nc
 import os
-from matplotlib.colors import ListedColormap
-import matplotlib.colors as mcolors
-import copy
-import itertools
-from cycler import cycler
-font_size = 13
-np.set_printoptions(suppress=True)
-plt.rcParams.update({"font.size": font_size})
+import shutil
 
-def flip(items, ncol):
-    return list(itertools.chain(*[items[i::ncol] for i in range(ncol)]))
+# Function to create directories and copy only .inp files from the source directory
+def create_directories_and_copy_files(brawl_type, source_dir, num_walkers, overlaps):
+    # Cores list for which directories will be created
+    windows = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16]
 
-def inner_mod(a,b):
-    res = a%b
-    return res if not res else res-b if a<0 else res
+    # Check if source directory exists
+    if not os.path.exists(source_dir):
+        print(f"Source directory '{source_dir}' does not exist.")
+        return
 
-colors = {
-    "steel_blue": "#1F77B4",
-    "light_steel_blue": "#AEC7E8",
-    "orange": "#FF7F0E",
-    "light_orange": "#FFBB78",
-    "forest_green": "#2CA02C",
-    "light_green": "#98DF8A",
-    "firebrick_red": "#D62728",
-    "soft_red": "#FF9896",
-    "lavender": "#9467BD",
-    "light_lavender": "#C5B0D5",
-    "brown": "#8C564B",
-    "tan": "#C49C94",
-    "orchid": "#E377C2",
-    "light_orchid": "#F7B6D2",
-    "gray": "#7F7F7F",
-    "light_gray": "#C7C7C7",
-    "yellow_green": "#BCBD22",
-    "light_yellow_green": "#DBDB8D",
-    "turquoise": "#17BECF",
-    "light_turquoise": "#9EDAE5"
-}
+    # Loop through the range of values for windows
+    for brawl in brawl_type:
+      for window in windows:
+        for walkers in num_walkers:
+          for overlap in overlaps:
+            for i in range(5):
+                # Create the directory name following the naming convention
+                dir_name = f'{brawl}_{walkers:02d}_{window:02d}_{overlap:02d}_{i+1}'
 
-custom_cmap = ListedColormap(colors.values())
-mpl.rcParams['axes.prop_cycle'] = mpl.cycler(color=custom_cmap.colors)
+                # Create the directory if it does not exist
+                if not os.path.exists(dir_name):
+                    os.makedirs(dir_name)
 
-filename = "data/wl_dos_bins.dat"
-bin_edges = nc.Dataset(filename)
-bin_edges = np.array(bin_edges["grid data"][:], dtype=np.float64)
+                for item in os.listdir(source_dir):
+                    source_item = os.path.join(source_dir, item)
+                    destination_item = os.path.join(dir_name, item)
+                    
+                    # Only copy .inp or .vij files, and skip directories
+                    if os.path.isfile(source_item) and item.endswith(('.inp', '.vij')):
[…]
