### 2025-08-05 e3e8ebd
Adjusted
