+++ examples/02_wang-landau_AlTiCrMo/wl_vis.py
@@ -14,33 +14,33 @@ np.set_printoptions(suppress=True)
 plt.rcParams.update({"font.size": font_size})
 
 def flip(items, ncol):
-    return list(itertools.chain(*[items[i::ncol] for i in range(ncol)]))
+	return list(itertools.chain(*[items[i::ncol] for i in range(ncol)]))
 
 def inner_mod(a,b):
-    res = a%b
-    return res if not res else res-b if a<0 else res
+	res = a%b
+	return res if not res else res-b if a<0 else res
 
 colors = {
-    "steel_blue": "#1F77B4",
-    "light_steel_blue": "#AEC7E8",
-    "orange": "#FF7F0E",
-    "light_orange": "#FFBB78",
-    "forest_green": "#2CA02C",
-    "light_green": "#98DF8A",
-    "firebrick_red": "#D62728",
-    "soft_red": "#FF9896",
-    "lavender": "#9467BD",
-    "light_lavender": "#C5B0D5",
-    "brown": "#8C564B",
-    "tan": "#C49C94",
-    "orchid": "#E377C2",
-    "light_orchid": "#F7B6D2",
-    "gray": "#7F7F7F",
-    "light_gray": "#C7C7C7",
-    "yellow_green": "#BCBD22",
-    "light_yellow_green": "#DBDB8D",
-    "turquoise": "#17BECF",
-    "light_turquoise": "#9EDAE5"
+	"steel_blue": "#1F77B4",
+	"light_steel_blue": "#AEC7E8",
+	"orange": "#FF7F0E",
+	"light_orange": "#FFBB78",
+	"forest_green": "#2CA02C",
+	"light_green": "#98DF8A",
+	"firebrick_red": "#D62728",
+	"soft_red": "#FF9896",
+	"lavender": "#9467BD",
+	"light_lavender": "#C5B0D5",
+	"brown": "#8C564B",
+	"tan": "#C49C94",
+	"orchid": "#E377C2",
+	"light_orchid": "#F7B6D2",
+	"gray": "#7F7F7F",
+	"light_gray": "#C7C7C7",
+	"yellow_green": "#BCBD22",
+	"light_yellow_green": "#DBDB8D",
+	"turquoise": "#17BECF",
+	"light_turquoise": "#9EDAE5"
 }
 
 custom_cmap = ListedColormap(colors.values())
@@ -63,12 +63,12 @@ U_data = rho_of_E.variables['U data'][:]
 
 elements = ""  # Initialize an empty string to store species data
 with open("brawl.inp", 'r') as file:
-    for line in file:
-        if 'species_name' in line:  # Check if the line contains the species_name variable
-            parts = line.split("=")  # Split the line by '=' and get everything after it
-            if len(parts) > 1:
-                elements = parts[1].strip().split()  # Get the data after '=' and strip spaces
-            break  # Exit the loop after finding species_name
+	for line in file:
+		if 'species_name' in line:  # Check if the line contains the species_name variable
+			parts = line.split("=")  # Split the line by '=' and get everything after it
+			if len(parts) > 1:
+				elements = parts[1].strip().split()  # Get the data after '=' and strip spaces
+			break  # Exit the loop after finding species_name
 n_species = len(elements)
 concentrations = 1.0/n_species*np.ones(n_species)
 
[…]
