### 2025-08-06 e394693
Adjusting parameters
