### 2025-08-06 90c3c41
Added bin count print out
