### 2025-08-06 9176888
Added overlap size print
