### 2025-08-06 f4158d6
Update generate_documentation.yml
