### 2025-08-06 6037ab4
Update generate_documentation.yml
