+++ .github/workflows/generate_documentation.yml
@@ -3,6 +3,12 @@ on:
   push:
     branches:
       - main
+
+permissions:
+  contents: read
+  pages: write
+  id-token: write
+      
 jobs:
   generate-documentation:
     runs-on: ubuntu-latest
