### 2025-08-06 1bca91d
Changed miminimum domain size back to 2% of total bins
