### 2025-08-06 213fc53
Adding information about viewing documentation hosted by GitHub Pages to README
