+++ README.md
@@ -4,8 +4,6 @@
 
 Copyright (C) The Authors 2019-2025. Released under the GNU Lesser General Public License, version 3.
 
----
-
 ## Background
 
 The Bragg-Williams Hamiltonian is an on-lattice Ising-like Hamiltonian describing the internal energy of a general substitutional alloy. The configuration of the alloy is specified by the *site occupation numbers*, $\{\xi_{i\alpha}\}$, where $\xi_{i\alpha}=1$ if site $i$ is occupied by an atom of species $\alpha$, and $\xi_{i\alpha}=0$ otherwise. (These can be thought of as a little like the 'spins' of the Ising model.) Each lattice site must be constrained to have one (and only one) atom sitting on it, expressed as
@@ -125,12 +123,15 @@ To save the output of the test run, pipe it to a file, _e.g._
 The output of this routine will tell you which tests (if any failed).
 
 ## Documentation
-The in addition to this README and the provided examples, the code also has (searchable) documentation which is auto-generated using [Doxygen](https://www.doxygen.nl), which lets users 'host' a web interface to the documentation locally on their machine. To view this documentation: 
+The in addition to this README and the provided examples, the code also has (searchable) documentation which is auto-generated using [Doxygen](https://www.doxygen.nl). This documentation contains information about all modules, functions, subroutines, and derived types, so is particularly useful if you are looking to develop a new feature. There are two options available for viewing this documentation:
+1. View the auto-generated documentation hosted online at [https://chriswoodgate.github.io/BraWl/](https://chriswoodgate.github.io/BraWl/). (Intended for end-users.)
+2. Generate and 'host' your own version of the documentation on your local machine. (Intended for developers wanting to check that their comments/documentation is being rendered as expected.)
+
+### Local generation and 'hosting' of documentation
+To generate and host your own version of the documentation, you should:
 1. Obtain [Doxygen](https://www.doxygen.nl), which is typically available through a package manager.
 2. Run `doxygen docs/doxyfile` from the code's main directory.
 3. Navigate to the freshly-generated `docs/html` directory and `open` index.html. (Alternatively, use your system's filemanager to open this in your web browser.)
-
-This documentation contains information about all modules, functions, subroutines, and derived types, so is particularly useful if you are looking to develop a new feature.
 
 ## Citations
 If you use `BraWl` in your research, please cite our preprint:
