### 2025-08-06 8c33af4
Removing trailing whitespace in comments for Doxygen
