### 2025-08-06 e9582e4
Tweaks to documentation mainpage for web viewing, and inclusion of relevant links
