### 2025-08-06 a836466
Adding some content to the documentation mainpage
