### 2025-08-06 b56d56d
Updated contributing information in the README
