### 2025-08-07 6fb42d5
Test to plot subplots
