+++ examples/plotting_scripts/performance_plotting.py
@@ -51,15 +51,14 @@ methods = np.array([0, 1, 2, 3, 4, 5])
 methods_label = np.array(["NU-LB-R", "NU-LB", "NU-R", "NU", "R", "Domains"])
 windows = np.array([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16])
 walkers = np.array([1, 2, 3, 4, 5, 6])
-overlaps = np.array([0, 10, 25, 50, 75])
+overlaps = np.array([5, 10, 25, 50, 75])
 repeats = 5
 bins=512
 
-walkers = np.array([1])
-#methods = np.array([0, 2, 4])
-#methods_label = np.array(["Non-Uniform + Balance + Replica", "Non-Uniform + Replica", "Replica"])
+methods = np.array([0, 2, 4])
+methods_label = np.array(["Non-Uniform + Balance + Replica", "Non-Uniform + Replica", "Replica"])
 #methods = np.array([0,1])
-#walkers = np.array([1])
+walkers = np.array([1])
 #overlaps = np.array([50])
 
 sums = np.zeros([len(methods), len(walkers), len(windows), len(overlaps), repeats])
@@ -81,9 +80,6 @@ for method_id, method in enumerate(methods):
           wl_lb_mc_steps = nc.Dataset(filename)
           wl_lb_mc_steps = np.array(wl_lb_mc_steps["grid data"][:], dtype=np.int64)
           mc_steps[method_id, walker, window, overlap, k-1] = np.sum(wl_lb_mc_steps)
-
-# REMOVE THIS LATER
-#sums[0, 5, 8, 0, 0] = 250
 
 one_sums = sums[:, :, 0, :, :]
 
@@ -122,28 +118,39 @@ efficiency_err = np.empty_like(sums_err)
 
 # Window Efficiency
 for overlap in range(len(overlaps)):
-  max_y = 0
   for method_id, method in enumerate(methods):
     for walker in range(len(walkers)):
       efficiency[method_id, walker, :, overlap] = one_sums_mean[walker, overlap]/(sums_mean[method_id, walker, :, overlap]*windows)
       efficiency_err[method_id, walker, :, overlap] = efficiency[method_id, walker, :, overlap]*np.sqrt((one_sums_err[walker, overlap]/one_sums_mean[walker, overlap])**2+(sums_err[method_id, walker, :, overlap]/sums_mean[method_id, walker, :, overlap])**2)
+
+selected_methods = [0, 2, 4]
+fix, axes = plt.subplots(1, 3, figsize=(18, 6), sharey=True)
+
+max_y = 0
+for overlap in range(len(overlaps)):
+  for method_id in selected_methods:
+    for walker in range(len(walkers)):
       if max_y < np.max(efficiency[method_id, walker, :, overlap]):
         max_y = np.max(efficiency[method_id, walker, :, overlap])
-  for method_id, method in enumerate(methods):
-    plt.axhline(y=1, linestyle='-')
+
+for ax_id, method_id in enumerate(selected_methods):
+    ax = axes[ax_id]
+    ax.axhline(y=1, linestyle='-')
+
     for walker in range(len(walkers)):
       plt.errorbar(windows, efficiency[method_id, walker, :, overlap], yerr=efficiency_err[method_id, walker, :, overlap], capsize=3, ls='none', fmt='o', label="{}".format(walker+1))
-    plt.xticks(windows, labels=windows)
-    plt.gca().set_box_aspect(1)
-    plt.xlabel("Domains")
-    plt.ylabel("Domain Efficiency")
[…]
