### 2025-08-07 e0188f9
Test to plot subplots
