### 2025-08-07 6a2416d
Test to plot subplots
