### 2025-08-08 854f50a
Updated for paper plot style
