+++ examples/plotting_scripts/performance_plotting.py
@@ -9,17 +9,134 @@ from matplotlib.colors import ListedColormap
 import matplotlib.colors as mcolors
 import itertools
 from scipy.optimize import curve_fit
+import matplotlib.ticker as ticker
+import matplotlib.ticker as mticker
 
-font_size = 16
+def plot_domain_efficiency(selected_methods, replica=True):
+  if replica:
+    replica = "r"
+  else:
+    replica = "nr"
+  
+  for overlap in range(len(overlaps)):
+    fig, axes = plt.subplots(1, 3, figsize=figsize_subplots)
+    max_y = 0
+    for method_id in selected_methods:
+      for walker in range(len(walkers)):
+        if max_y < np.max(efficiency[method_id, walker, :, overlap]):
+          max_y = np.max(efficiency[method_id, walker, :, overlap])
+
+    for ax_id, method_id in enumerate(selected_methods):
+      ax = axes[ax_id]
+      ax.axhline(y=1, linestyle='-')
+
+      for walker in range(len(walkers)):
+        ax.errorbar(windows, efficiency[method_id, walker, :, overlap], yerr=efficiency_err[method_id, walker, :, overlap], capsize=3, ls='none', fmt='o', label="{}".format(walker+1))
+      ax.set_xticks(windows[::2])
+      ax.set_xlabel("Domains")
+      ax.tick_params(axis="y", direction='inout')
+      if ax_id == 0:
+        ax.set_ylabel("Domain Efficiency")
+        ax.yaxis.set_major_formatter(ticker.FormatStrFormatter('%.1f'))
+      else:
+        ax.set_yticklabels([])
+      #plt.title("Method {}".format(methods_label[method_id]))
+      #plt.legend(loc="upper right", title="Walkers")
+      ax.set_ylim(0, max_y*1.25)
+      ax.text(0.5, -0.225, titles[ax_id], transform=ax.transAxes, ha='center', va='top')
+    
+    
+    plt.subplots_adjust(wspace=0, hspace=0)
+    plt.savefig('{:02d}_{}_domain_efficiency.pdf'.format(overlaps[overlap], replica), bbox_inches='tight')
+    plt.close()
+
+def plot_mc_step(selected_methods, replica=True):
+  if replica:
+    replica = "r"
+  else:
+    replica = "nr"
+  
+  for overlap in range(len(overlaps)):
+    fig, axes = plt.subplots(1, 3, figsize=figsize_subplots)
+    max_y = 0
+    for method_id in selected_methods:
+      for walker in range(len(walkers)):
+        if max_y < np.max(mc_steps_mean[method_id, walker, :, overlap]):
+          max_y = np.max(mc_steps_mean[method_id, walker, :, overlap])
+
+    for ax_id, method_id in enumerate(selected_methods):
+      ax = axes[ax_id]
+      ax.axhline(y=1, linestyle='-')
+
+      for walker in range(len(walkers)):
+        ax.errorbar(windows, mc_steps_mean[method_id, walker, :, overlap], yerr=mc_steps_err[method_id, walker, :, overlap], capsize=3, ls='none', fmt='o', label="{}".format(methods_label[method_id]))
+      ax.set_xticks(windows[::2])
+      ax.set_xlabel("Cores")
+      ax.tick_params(axis="y", direction='inout')
+      if ax_id == 0:
+        ax.set_ylabel("Relative MC Steps")
+        ax.yaxis.set_major_formatter(ticker.FormatStrFormatter('%.1f'))
+      else:
+        ax.set_yticklabels([])
[…]
