### 2025-08-11 ed8aec7
Added inifinite well boundary condition to sweeps
