### 2025-08-19 c7f72ad
Error-handling of (small) negative heat capacity due to rounding errors
