### 2025-08-21 c5b75a4
Updated publications list in README.md
