### 2025-08-21 53aac6c
Updating publications list in documentation mainpage
