### 2025-08-21 8d69ec4
Adjustment of contributors list in README
