### 2025-08-21 6d9b21a
Tweaks to Makefile for MacOS/Homebrew builds
