### 2025-08-21 7658baf
Addition of writing of initial state (used in testing routines)
