### 2025-08-21 7694f66
Switching to reinstall of gcc
