### 2025-08-21 a03d77c
Adding button for MacOS tests (ARM64) to README
