### 2025-08-21 f364b1e
Renaming Apple workflow to reflect ARM64 architecture
