+++ .github/workflows/testingOnPush_Apple.yaml
@@ -1,4 +1,4 @@
-name: MacOS (M1)
+name: MacOS (ARM64)
 
 on:
   push:
