### 2025-08-21 e72a06a
Adding ability to write initial grid configurations at each temperature in Metropolis routines
