+++ .github/workflows/testingOnPush_Ubuntu.yaml
@@ -55,3 +55,10 @@ jobs:
           make compiler=mpifort tests
           cd tests
           mpirun -np 4 ../tests.run
+          
+      - name: Run WL test with mpifort with 4 processes
+        run: |
+          make compiler=mpifort tests
+          cd tests/04_parallel_wang-landau
+          mpirun -np 4 ../../brawl.run
+          python wl_test.py
