### 2025-08-22 2b771ec
Tweaking static seed used for testing
