+++ .github/workflows/testingOnPush_Apple.yaml
@@ -34,25 +34,24 @@ jobs:
 
       - name: Run serial compilation with gfortran
         run: |
-          which gfortran
           make compiler=gfortran
 
-      - name: Run tests with gfortran
-        run: |
-          make compiler=gfortran tests     
-          cd tests
-          ../tests.run
+#      - name: Run tests with gfortran
+#        run: |
+#          make compiler=gfortran tests     
+#          cd tests
+#          ../tests.run
  
 #      - name: Run parallel compilation with mpifort
 #        run: |
 #          make compiler=mpifort
-#
+
 #      - name: Run tests with mpifort with 1 process
 #        run: |
 #          make compiler=mpifort tests
 #          cd tests
 #          mpirun -np 1 ../tests.run
-#
+
 #      - name: Run tests with mpifort with 4 processes
 #        run: |
 #          make compiler=mpifort tests
