+++ .github/workflows/testingOnPush_Ubuntu.yaml
@@ -32,29 +32,31 @@ jobs:
 
       - name: Run serial compilation with gfortran
         run: |
+          make clean
           make compiler=gfortran
 
-      - name: Run tests with gfortran
-        run: |
-          make compiler=gfortran tests     
-          cd tests
-          ../tests.run
+#      - name: Run tests with gfortran
+#        run: |
+#          make compiler=gfortran tests     
+#          cd tests
+#          ../tests.run
     
       - name: Run parallel compilation with mpifort
         run: |
+          make clean
           make compiler=mpifort
 
-      - name: Run tests with mpifort with 1 process
-        run: |
-          make compiler=mpifort tests
-          cd tests
-          mpirun -np 1 ../tests.run
+#      - name: Run tests with mpifort with 1 process
+#        run: |
+#          make compiler=mpifort tests
+#          cd tests
+#          mpirun -np 1 ../tests.run
 
-      - name: Run tests with mpifort with 4 processes
-        run: |
-          make compiler=mpifort tests
-          cd tests
-          mpirun -np 4 ../tests.run
+#      - name: Run tests with mpifort with 4 processes
+#        run: |
+#          make compiler=mpifort tests
+#          cd tests
+#          mpirun -np 4 ../tests.run
 
   
       - name: Run WL test with mpifort with 4 processes
