+++ paper/brawl.bib
@@ -337,14 +337,14 @@
 
 @article{woodgateemergent2025,
 	title = {Emergent {B2} chemical orderings in the {AlTiVNb} and {AlTiCrMo} refractory high-entropy superalloys studied via first-principles theory and atomistic modelling},
-	copyright = {arXiv.org perpetual, non-exclusive license},
-	url = {https://doi.org/10.48550/arXiv.2503.13235},
-	urldate = {2025-03-25},
-	journal = {arXiv:2503.13235},
-	author = {Woodgate, Christopher D. and Naguszewski, Hubert J. and Redka, David and Minár, Ján and Quigley, David and Staunton, Julie B.},
-    year = {2025},
-    doi = {10.48550/arXiv.2503.13235},
-	keywords = {Condensed Matter - Materials Science, Physics - Applied Physics, Physics - Computational Physics},
+	volume = {8},
+	url = {https://doi.org/10.1088/2515-7639/adf468},
+	doi = {10.1088/2515-7639/adf468},
+	journal = {Journal of Physics: Materials},
+	author = {Woodgate, Christopher D. and Naguszewski, Hubert J. and Redka, David and Ján, Minár and David, Quigley and Staunton, Julie B.},
+	month = aug,
+	year = {2025},
+	pages = {045002},
 }
 
 @article{ducastellegeneralized1976,
