### 2025-08-22 3f38b55
Correction to author list of JPhys reference, and also removal of some incorrect issue numbers
