+++ paper/brawl.bib
@@ -21,7 +21,6 @@
 	issn = {2469-9950, 2469-9969},
 	url = {https://doi.org/10.1103/PhysRevB.105.115124},
 	doi = {10.1103/PhysRevB.105.115124},
-	number = {11},
 	urldate = {2022-03-25},
 	journal = {Physical Review B},
 	author = {Woodgate, Christopher D. and Staunton, Julie B.},
@@ -36,7 +35,6 @@
 	issn = {0021-8979},
 	url = {https://doi.org/10.1063/5.0200862},
 	doi = {10.1063/5.0200862},
-	number = {13},
 	journal = {Journal of Applied Physics},
 	author = {Woodgate, Christopher D. and Staunton, Julie B.},
 	month = apr,
@@ -105,7 +103,6 @@
 	shorttitle = {Short-range order and compositional phase stability in refractory high-entropy alloys via first-principles theory and atomistic modeling},
 	url = {https://doi.org/10.1103/PhysRevMaterials.7.013801},
 	doi = {10.1103/PhysRevMaterials.7.013801},
-	number = {1},
 	urldate = {2023-02-02},
 	journal = {Physical Review Materials},
 	author = {Woodgate, Christopher D. and Staunton, Julie B.},
@@ -121,7 +118,6 @@
 	shorttitle = {Interplay between magnetism and short-range order in medium- and high-entropy alloys},
 	url = {https://doi.org/10.1103/PhysRevMaterials.7.053801},
 	doi = {10.1103/PhysRevMaterials.7.053801},
-	number = {5},
 	urldate = {2023-05-09},
 	journal = {Physical Review Materials},
 	author = {Woodgate, Christopher D. and Hedlund, Daniel and Lewis, L. H. and Staunton, Julie B.},
@@ -223,9 +219,8 @@
 	shorttitle = {Structure, short-range order, and phase stability of the {AlxCrFeCoNi} high-entropy alloy},
 	url = {https://doi.org/10.1038/s41524-024-01445-w},
 	doi = {10.1038/s41524-024-01445-w},
-	number = {1},
 	urldate = {2024-12-08},
-	journal = {npj Computational Materials},
+	journal = {{npj} Computational Materials},
 	author = {Woodgate, Christopher D. and Marchant, George A. and Pártay, Livia B. and Staunton, Julie B.},
 	month = nov,
 	year = {2024},
@@ -238,9 +233,8 @@
 	issn = {2057-3960},
 	url = {https://doi.org/10.1038/s41524-024-01435-y},
 	doi = {10.1038/s41524-024-01435-y},
-	number = {1},
 	urldate = {2024-12-08},
-	journal = {npj Computational Materials},
+	journal = {{npj} Computational Materials},
 	author = {Woodgate, Christopher D. and Lewis, Laura H. and Staunton, Julie B.},
 	month = nov,
 	year = {2024},
@@ -341,7 +335,7 @@
 	url = {https://doi.org/10.1088/2515-7639/adf468},
 	doi = {10.1088/2515-7639/adf468},
 	journal = {Journal of Physics: Materials},
-	author = {Woodgate, Christopher D. and Naguszewski, Hubert J. and Redka, David and Ján, Minár and David, Quigley and Staunton, Julie B.},
+	author = {Woodgate, Christopher D. and Naguszewski, Hubert J. and Redka, David and Minár, Ján and Quigley, David and Staunton, Julie B.},
 	month = aug,
 	year = {2025},
 	pages = {045002},
