### 2025-08-23 b3c2873
Created for CI tests
