### 2025-08-23 44cb699
Changed netCDF output file names to .nc from .dat
