+++ .github/workflows/testingOnPush_Ubuntu.yaml
@@ -34,35 +34,32 @@ jobs:
         run: |
           make clean
           make compiler=gfortran
+      
+      - name: Run serial Metropolis test
+        run: |
+          cd test/01_serial_metropolis
+          ../../brawl.run
+          python ../ci_test.py
 
-#      - name: Run tests with gfortran
-#        run: |
-#          make compiler=gfortran tests     
-#          cd tests
-#          ../tests.run
-    
+      - name: Run serial Nested Sampling test
+        run: |
+          cd test/03_serial_nested_sampling
+          ../../brawl.run
+          python ../ci_test.py
+
       - name: Run parallel compilation with mpifort
         run: |
           make clean
           make compiler=mpifort
 
-#      - name: Run tests with mpifort with 1 process
-#        run: |
-#          make compiler=mpifort tests
-#          cd tests
-#          mpirun -np 1 ../tests.run
-
-#      - name: Run tests with mpifort with 4 processes
-#        run: |
-#          make compiler=mpifort tests
-#          cd tests
-#          mpirun -np 4 ../tests.run
-
-  
-      - name: Run WL test with mpifort with 4 processes
+      - name: Run prallel Metropolis test
         run: |
-          make compiler=mpifort clean
-          make compiler=mpifort
-          cd tests/04_parallel_wang-landau
+          cd test/03_parallel_metropolis
           mpirun -np 4 ../../brawl.run
-          python wl_test.py
+          python ../ci_test.py
+
+      - name: Run parallel Wang-Landau test
+        run: |
+          cd test/04_parallel_wang-landau
+          mpirun -np 4 ../../brawl.run
+          python ../ci_test.py
