### 2025-08-23 39d8d3b
Changed nc test to be NRMSE (less than 1%)
