### 2025-08-23 3ed09a8
Added special case for WL to use NRMSE and exact match for all others
