+++ tests/ci_test.py
@@ -38,14 +38,20 @@ for i, _ in enumerate(ref_files):
           ref_var = ref_nc.variables[var_name][:]
           test_var = test_nc.variables[var_name][:]
 
-          ref_flat = ref_var.flatten()
-          test_flat = test_var.flatten()
+          # if cwd is 04_parallel_wang-landau
+          if cwd == "04_parallel_wang-landau":
+            ref_flat = ref_var.flatten()
+            test_flat = test_var.flatten()
 
-          nrmse = np.sqrt(np.mean((ref_flat - test_flat) ** 2))/np.mean(np.abs(ref_flat))
+            nrmse = np.sqrt(np.mean((ref_flat - test_flat) ** 2))/np.mean(np.abs(ref_flat))
 
-          if nrmse > 0.01:
-            print(f"[DATA MISMATCH NRMSE >1%] {os.path.basename(test_files[i])}, variable: {var_name}")
-            sys.exit(1)
+            if nrmse > 0.01:
+              print(f"[DATA MISMATCH NRMSE >1%] {os.path.basename(test_files[i])}, variable: {var_name}")
+              sys.exit(1)
+          else:
+            if not np.array_equal(ref_var, test_var):
+              print(f"[DATA MISMATCH] {os.path.basename(test_files[i])}, variable: {var_name}")
+              sys.exit(1)
 
     except Exception as e:
       print(f"[FILE ERROR] {file_name}: {e}")
