### 2025-09-01 849d507
Adding full test suite for MacOS CI
