### 2025-09-01 6e9fa06
Forcing oversubscribe on parallel threads for testing of parallel routines
