### 2025-09-01 4ee8429
Removing old (no longer needed) testing routines
