### 2025-09-01 86c5702
Tweaks to README to improve discussion of examples.
