+++ paper/brawl.bib
@@ -21,6 +21,7 @@
 	issn = {2469-9950, 2469-9969},
 	url = {https://doi.org/10.1103/PhysRevB.105.115124},
 	doi = {10.1103/PhysRevB.105.115124},
+	number = {11},
 	urldate = {2022-03-25},
 	journal = {Physical Review B},
 	author = {Woodgate, Christopher D. and Staunton, Julie B.},
@@ -35,6 +36,7 @@
 	issn = {0021-8979},
 	url = {https://doi.org/10.1063/5.0200862},
 	doi = {10.1063/5.0200862},
+	number = {13},
 	journal = {Journal of Applied Physics},
 	author = {Woodgate, Christopher D. and Staunton, Julie B.},
 	month = apr,
@@ -103,6 +105,7 @@
 	shorttitle = {Short-range order and compositional phase stability in refractory high-entropy alloys via first-principles theory and atomistic modeling},
 	url = {https://doi.org/10.1103/PhysRevMaterials.7.013801},
 	doi = {10.1103/PhysRevMaterials.7.013801},
+	number = {1},
 	urldate = {2023-02-02},
 	journal = {Physical Review Materials},
 	author = {Woodgate, Christopher D. and Staunton, Julie B.},
@@ -118,6 +121,7 @@
 	shorttitle = {Interplay between magnetism and short-range order in medium- and high-entropy alloys},
 	url = {https://doi.org/10.1103/PhysRevMaterials.7.053801},
 	doi = {10.1103/PhysRevMaterials.7.053801},
+	number = {5},
 	urldate = {2023-05-09},
 	journal = {Physical Review Materials},
 	author = {Woodgate, Christopher D. and Hedlund, Daniel and Lewis, L. H. and Staunton, Julie B.},
@@ -219,6 +223,7 @@
 	shorttitle = {Structure, short-range order, and phase stability of the {AlxCrFeCoNi} high-entropy alloy},
 	url = {https://doi.org/10.1038/s41524-024-01445-w},
 	doi = {10.1038/s41524-024-01445-w},
+	number = {1},
 	urldate = {2024-12-08},
 	journal = {{npj} Computational Materials},
 	author = {Woodgate, Christopher D. and Marchant, George A. and Pártay, Livia B. and Staunton, Julie B.},
@@ -233,6 +238,7 @@
 	issn = {2057-3960},
 	url = {https://doi.org/10.1038/s41524-024-01435-y},
 	doi = {10.1038/s41524-024-01435-y},
+	number = {1},
 	urldate = {2024-12-08},
 	journal = {{npj} Computational Materials},
 	author = {Woodgate, Christopher D. and Lewis, Laura H. and Staunton, Julie B.},
