### 2025-09-01 #8346 @ChrisWoodgate
Hi @jgostick,

Sorry for the delay on this!

@HubertJN and I have now addressed all the testing/CI issues raised by @sudarshanv01 and @amkrajewski, and I have marked the relevant issues as 'closed' on the code repository.

Looking at the reviewer checklists, I think the main items which will remain outstanding are those relating to @amkrajewski reviewing the paper itself.

Otherwise, it looks like we are mostly there!
