### 2025-09-02 #8346 @ChrisWoodgate
For reference, @amkrajewski, you can find our as-submitted text (in Physical Review style) on the arXiv: [https://doi.org/10.48550/arXiv.2505.05393](https://doi.org/10.48550/arXiv.2505.05393).

It might be worth you taking a look at this and see if at least some of the issues which you have raised are addressed there. 🙂
