### 2025-09-10 #8346 @ChrisWoodgate
Hi @jgostick,

Okay, thanks for the clarification.

@amkrajewski, could you please give us some guidance on where you think the additional content you would like included in the paper might be able to go? (If we include it in the paper itself, it looks as though we are going to fall foul of the broader editorial guidance.)

Additionally, regarding this point:

> It would be good to include some brief performance metrics and note on how the cost scales with the system size.

Could you clarify what sort of performance metrics you would like included, please? Are you looking for something simple like, _e.g._, how many Metropolis-Hastings trial swaps the code achieves for a particular system on a particular architecture in a given time frame, or something more complex? (Scaling is easier to address as, due to the finite cutoff of pair-pair interactions, the computational cost of evaluating the Hamiltonian scales linearly with the number of atoms present in the system for sufficiently large systems.)

Thanks both!

P.S. Apologies for the slow response—I am currently out of the office on holiday. (Most universities in the UK don't get back up and running until mid- to late-September.)
