### 2025-09-10 02b8020
Minor tweaks to README including updated publications list.
