### 2025-09-29 #8346 @ChrisWoodgate
Hi @jgostick, @amkrajewski,

Could we perhaps try to move this forward this week, please? (This review was originally on June 4, which is now almost four months ago.)

It looks like the software portion of the review is now complete, and all we need to do is find a consensus on where the additional material which @amkrajewski would like included in the paper can go, as well as a couple of clarifications on some of the details to be included in this material.

I appreciate that we are all busy at the moment, but if we get this finished (which shouldn't take long), we could all tick this particular item off of our 'to do' lists!

Thanks!
