### 2025-09-30 #8346 @ChrisWoodgate
Thanks for the clarifications, @amkrajewski! I appreciate you taking the time; it sounds like you are very busy at the moment.

Okay, I think I have a way by which we can proceed here which should be acceptable to everybody. I think we will aim to put most of the requested additional content in a revised arXiv preprint, linked from the documentation. The one exception will probably be for the performance metrics, which we will put in the JOSS paper as this feels highly relevant to that. I will also fix the problem-child reference—thanks again for pointing this out.

Leave it with me and @HubertJN... 🚀
