### 2025-10-06 4458ef3
Updated plotting scripts
