### 2025-10-06 d4cc7f1
Changed minimum bins to 1%
