+++ src/wang-landau.F90
@@ -1223,7 +1223,7 @@ module wang_landau
         bins = NINT(REAL(wl_setup_internal%bins)*diffusion_merge/SUM(diffusion_merge))
         
         ! Set all bins less than min_bins to min_bins
-        min_bins = MAX(INT(0.02_real64*wl_setup_internal%bins), 2)
+        min_bins = MAX(INT(0.01_real64*wl_setup_internal%bins), 2)
         do i = 1, wl_setup_internal%num_windows
           if (bins(i) < min_bins) then
               bins(i) = min_bins
