### 2025-10-06 781d0eb
Minor tweak to README
