+++ src/wang-landau.F90
@@ -28,52 +28,51 @@ module wang_landau
   use display
 
   implicit none
- 
+
   ! MPI variables
   integer :: mpi_processes, ierr
-  real(real64) :: start, end, time_max, time_min, test_time_1, test_time_2
   integer :: mpi_bins, mpi_start_idx, mpi_end_idx, mpi_index, mpi_start_idx_buffer
-  integer :: mpi_end_idx_buffer, i_sweeps
-  real(real64) :: scale_factor, scale_count, wl_logdos_min, bin_overlap, beta_diff, beta_original, beta_merge
-  real(real64), allocatable :: mpi_bin_edges(:), mpi_wl_hist(:), wl_logdos_buffer(:), wl_logdos_combine(:), window_overlap(:,:)
-  real(real64), allocatable :: rank_time(:), rank_time_buffer(:,:)
- 
+  integer :: mpi_end_idx_buffer, i_sweeps, num_walkers
+  real(real64) :: start, end, time_max, time_min, test_time_1, test_time_2
+  real(real64) :: scale_factor, scale_count, wl_logdos_min, bin_overlap
+  real(real64) :: beta_diff, beta_original, beta_merge
+  real(real64), allocatable :: mpi_bin_edges(:), mpi_wl_hist(:)
+  real(real64), allocatable :: wl_logdos_buffer(:), wl_logdos_combine(:)
+  real(real64), allocatable :: window_overlap(:,:), rank_time(:), rank_time_buffer(:,:)
+
   ! Window variables
-  integer, allocatable :: window_intervals(:,:), window_rank_index(:,:)
-  integer, allocatable :: window_indices(:, :)
-  integer :: num_walkers
-  
+  integer, allocatable :: window_intervals(:,:), window_rank_index(:,:), window_indices(:,:)
   ! Setup types
   type(run_params) :: setup_internal
   type(wl_params) :: wl_setup_internal
-  
+
   ! Temperature and temperature steps
   real(real64) :: temp, beta
-  
-  ! WL variables and arrays
+
+  ! Wang-Landau variables and arrays
   integer :: bins, resets, pre_sampled_state, ibin, itemp
   real(real64) :: bin_width, energy_to_ry, wl_f, wl_f_prev, tolerance, flatness_tolerance
   real(real64) :: target_energy, flatness, bins_buffer, bins_min, radial_min, radial_min_buffer
-  real(real64), allocatable :: bin_edges(:), wl_hist(:), wl_logdos(:), bin_energy(:), mean_energy(:,:), prob(:)
+  real(real64), allocatable :: bin_edges(:), wl_hist(:), wl_logdos(:)
+  real(real64), allocatable :: bin_energy(:), mean_energy(:,:), prob(:)
   integer, allocatable :: pre_sampled(:), pre_sampled_buffer(:)
   logical :: rho_saved
-  
   ! Load balancing metrics
-  real(real64), allocatable :: lb_bins(:,:), lb_avg_time(:,:), lb_max_time(:,:), lb_mc_steps(:), lb_mc_steps_buffer(:)
+  real(real64), allocatable :: lb_bins(:,:), lb_avg_time(:,:), lb_max_time(:,:)
+  real(real64), allocatable :: lb_mc_steps(:), lb_mc_steps_buffer(:)
   real(real64), allocatable :: window_time(:), diffusion_prev(:)
   integer :: converged, converged_sum
-  
+
   ! Radial density across energy
   real(real64), allocatable :: rho_of_E(:,:,:,:), rho_of_E_buffer(:,:,:,:)
   integer, allocatable :: radial_record(:), radial_record_buffer(:)
   logical, allocatable :: radial_record_bool(:)
   integer :: radial_mc_steps
[…]
