### 2025-10-14 5f8d658
Reduced number of pre-sample visists
