### 2025-10-16 9ffe5e7
implemented config loading for entering windows
