+++ src/wang-landau.F90
@@ -57,6 +57,8 @@ module wang_landau
   real(real64), allocatable :: bin_edges(:), wl_hist(:), wl_logdos(:), bin_energy(:), mean_energy(:,:), prob(:)
   integer, allocatable :: pre_sampled(:), pre_sampled_buffer(:)
   logical :: rho_saved
+  integer(array_int), allocatable :: config_per_bin(:,:,:,:,:)
+  logical, allocatable :: config_store(:)
   
   ! Load balancing metrics
   real(real64), allocatable :: lb_bins(:,:), lb_avg_time(:,:), lb_max_time(:,:), lb_mc_steps(:), lb_mc_steps_buffer(:)
@@ -273,7 +275,8 @@ module wang_landau
 
         call compute_mean_energy(wl_logdos)
 
-        call enter_energy_window()
+        !call enter_energy_window()
+        call load_window_config()
 
         call zero_subtract_logdos(wl_logdos)
         call comms_wait()
@@ -584,6 +587,10 @@ module wang_landau
         end if
         if (MOD(i,INT(0.02_real64*REAL(setup_internal%n_atoms))) == 0) then
           mpi_wl_hist(jbin - mpi_start_idx + 1) = mpi_wl_hist(jbin - mpi_start_idx + 1) + 1.0_real64
+          if (config_store(jbin)) then
+            config_store(jbin) = .False.
+            config_per_bin(jbin, :, :, :, :) = config
+          end if
         end if
         wl_logdos(jbin) = wl_logdos(jbin) + wl_f
       else
@@ -592,6 +599,10 @@ module wang_landau
         call pair_swap(config, rdm1, rdm2)
         if (MOD(i,INT(0.02_real64*REAL(setup_internal%n_atoms))) == 0) then
           mpi_wl_hist(jbin - mpi_start_idx + 1) = mpi_wl_hist(jbin - mpi_start_idx + 1) + 1.0_real64
+          if (config_store(jbin)) then
+            config_store(jbin) = .False.
+            config_per_bin(jbin, :, :, :, :) = config
+          end if
         end if
         wl_logdos(jbin) = wl_logdos(jbin) + wl_f
       end if
@@ -730,7 +741,6 @@ module wang_landau
   !> @date    2024 
   subroutine pre_sampling(wl_logdos, mpi_wl_hist)
     real(real64), allocatable, intent(inout) :: wl_logdos(:), mpi_wl_hist(:)
-
     i_sweeps = 0
     pre_sampled = 0
     pre_sampled_buffer = 0
@@ -788,7 +798,10 @@ module wang_landau
       write (*, *)
     end if
 
-    call enter_energy_window()
+    call merge_configs()
+
+    !call enter_energy_window()
+    call load_window_config()
 
     call zero_subtract_logdos(wl_logdos)
   end subroutine pre_sampling
@@ -1007,6 +1020,8 @@ module wang_landau
     allocate(bin_energy(bins))
     allocate(mean_energy(300,2))
     allocate(prob(bins))
+    allocate(config_per_bin(bins, setup_internal%n_basis, 2*setup_internal%n_1, 2*setup_internal%n_2, 2*setup_internal%n_3))
+    allocate(config_store(bins))
   end subroutine primary_array_allocation
 
   !> @brief   Second set of array allocations
@@ -1071,6 +1086,8 @@ module wang_landau
     radial_mc_steps = 0
     rho_saved = .False.
     lb_mc_steps_buffer = 0.0_real64; lb_mc_steps = 0.0_real64
+    config_per_bin = 0_array_int
+    config_store = .True.
 
     mean_energy = 1.0_real64 / (k_b_in_Ry * 10.0_real64)
 
@@ -1488,6 +1505,80 @@ module wang_landau
     rdm1 = setup_internal%rdm_site()
[…]
