### 2025-10-20 39466b8
Added basic shared arrays between walkers in the same window
