### 2025-10-20 #8346 @ChrisWoodgate
Hi @rkurchin,

Thanks for checking in on this. 🙂

This review process has been a long one, mainly due to misaligned author/reviewer availabilities. (It took almost eight weeks for a complete first round of reviews to come in, by which time I was busy with other commitments, which in turn delayed my response to the reviewers, which then impacted their availability and availability to respond, _etc._.)

It looks like the software portion of the review is now complete and all we have left to agree upon is the paper itself. The problem there is that both we as authors and @amkrajewski as a reviewer feels that there ought to be additional content in the paper, but we are already over the journal length limit and so are a bit 'stuck' with how to proceed.

**The solution we have arrived at is to have an elongated version of the paper uploaded to the arXiv that the formal JOSS paper refers to.** Personally, I think this is a bit daft; surely if both authors and reviewers want additional content in a paper then JOSS could make an exception? (To be honest, though, I just want this off my desk at this point; I feel like I have put more effort into this review process than for any other review process with any other journal! 😅)

Anyway, I have now written the updated version of the paper ready for arXiv upload and shared it with all co-authors for approval. This should hopefully go on to the arXiv next week, as an update to[ our original arXiv submission](https://doi.org/10.48550/arXiv.2505.05393), at which point I should be able to return an updated paper here.

Thoughts/comments/suggestions on any/all of the above welcome.

Thanks!
