### 2025-10-23 #8346 @jgostick
@ChrisWoodgate, I know how you feel, since you want a nice thorough paper to accompany your software release, but one of the reasons for insisting on short papers is so that the reviewers don't have to put their stamp of approval on a long winded, theory heavy paper, and can just focus on the software itself. So let's stick with that.

EDIT: I just realized that this statement is silly since it was the reviewer who asked for this extra info.

Anyway, this is just my personal feeling, but I don't love your plan to submit to arXiv. I worry that having an arXiv paper as well as a JOSS paper will split your citations. Why not just include this information in the documentation? This is entirely up to you of course.

I will await your uploading of the shortened JOSS paper, then we can proceed to the next step of the process.
