### 2025-10-23 #8346 @ChrisWoodgate
Hi @jgostick,

Sorry to be causing a headache for everybody here! 😅

> [@ChrisWoodgate](https://github.com/ChrisWoodgate), I know how you feel, since you want a nice thorough paper to accompany your software release, but one of the reasons for insisting on short papers is so that the reviewers don't have to put their stamp of approval on a long winded, theory heavy paper, and can just focus on the software itself. So let's stick with that.
>
> EDIT: I just realized that this statement is silly since it was the reviewer who asked for this extra info.

Yes, I think you have summarised my position fairly well. My _personal_ preference would be to submit a longer, self-contained paper which summarised the key theoretical background before describing the software, which was broadly the paper we submitted originally. (We also uploaded a version of this original document to the arXiv at the time of submission, as is standard practice in the computational physics/chemistry/materials community.) In the pre-review stage we cut the paper down substantially in response to editorial feedback, taking out the theoretical background and most of the discussion of the implemented methods, thus bringing it in line with JOSS guidance. Now, in the review stage, a reviewer has asked for additional content to be added, some of which is actually quite similar to content we took out at the pre-review stage.

Personally, I agree with @amkrajewski that the additional content and adjustments he has suggested would improve the paper, but I'm kind of stuck with where to put this additional content as we're already up against the length limit—more on this below.

> Anyway, this is just my personal feeling, but I don't love your plan to submit to arXiv. I worry that having an arXiv paper as well as a JOSS paper will split your citations. Why not just include this information in the documentation? This is entirely up to you of course.

I don't love it either, and usually the way I would handle this kind of situation (for a Letters journal or similar, with a length limit) would be to prepare some kind of 'supplementary material' in which any additional requested content/technical details could go. My understanding, though, is that JOSS doesn't permit supplementary material. Personally, I don't feel this additional content belongs in the documentation, either, because it's nothing to do with the usage/development of the software. So, the arXiv route seems to be a solution, even if an imperfect one.

I am not worried about split citations as this is already happening—another group cited our arXiv preprint for some of its methodological discussion just last week! (Amusingly, this is a citation we probably _wouldn't_ have received on the JOSS article in its current form as it related to a technical detail of one of the implementations.)

> I will await your uploading of the shortened JOSS paper, then we can proceed to the next step of the process.

Okay, thanks! 🙂

The updated JOSS paper should hopefully appear here next week some time—I am just waiting on co-authors to get back to me to finalise/approve the wording. (Most people in the UK are swamped with teaching at the moment, which is why we are moving quite slowly, sorry!)

EDIT: Happy to hear @rkurchin's thoughts on length, too. I actually thought that our original paper ([arXiv version here](https://doi.org/10.48550/arXiv.2505.05393)) was fairly short for a computational physics/chemistry paper, albeit a fair bit longer than JOSS guidelines.
