### 2025-10-30 6bf6fa4
Energy range explorer WIP
