### 2025-11-01 8ca9b2a
Adjusted window scaling and made min scaling value 10%
