### 2025-11-01 10c6b13
Updated load balance algorithm to function off of log DOS and mc steps
