+++ src/wang-landau.F90
@@ -54,7 +54,7 @@ module wang_landau
   integer :: bins, resets, pre_sampled_state, ibin, itemp
   real(real64) :: bin_width, wl_f, wl_f_prev, tolerance, flatness_tolerance
   real(real64) :: target_energy, flatness, bins_buffer, bins_min, radial_min, radial_min_buffer
-  real(real64), allocatable :: bin_edges(:), wl_hist(:), wl_logdos(:), bin_energy(:), mean_energy(:,:), prob(:)
+  real(real64), allocatable :: bin_edges(:), wl_hist(:), wl_logdos(:), bin_energy(:), mean_energy(:,:), prob(:), wl_mc_steps(:)
   integer, allocatable :: pre_sampled(:), pre_sampled_buffer(:)
   logical :: rho_saved
   integer(array_int), allocatable :: config_per_bin(:,:,:,:,:)
@@ -212,6 +212,9 @@ module wang_landau
       call sweeps(wl_logdos, mpi_wl_hist)
 
       lb_mc_steps(iter+1) = lb_mc_steps(iter+1) + wl_setup_internal%mc_sweeps*setup_internal%n_atoms
+      if (converged == 0) then
+        wl_mc_steps(mpi_index) = wl_mc_steps(mpi_index) + wl_setup_internal%mc_sweeps*setup_internal%n_atoms
+      end if
 
       if (MOD(i_sweeps, 10) == 0) then
         call replica_exchange(config)
@@ -244,6 +247,8 @@ module wang_landau
         call dos_average(wl_logdos)
         call dos_combine(wl_logdos)
 
+        call MPI_ALLREDUCE(MPI_IN_PLACE, wl_mc_steps, SIZE(wl_mc_steps), MPI_DOUBLE_PRECISION, MPI_SUM, MPI_COMM_WORLD, ierr)
+
         call reduce_time(start, end, radial_time)
 
         call comms_wait()
@@ -267,8 +272,11 @@ module wang_landau
                   " | Radial samples: ", radial_min*100_real64, "%"
             write (*, *)
           do i=1, wl_setup_internal%num_windows
-            write (6, '(a,i3,a,f12.2,a,f12.2,a,f12.2,a)') "MPI Window: ", i, " | Avg. time: ", rank_time_buffer(i,1), &
-            "s | Time min: ", rank_time_buffer(i,2), "s Time max: " , rank_time_buffer(i,3), "s"
+            !write (6, '(a,i3,a,f12.2,a,f12.2,a,f12.2,a)') "MPI Window: ", i, " | Avg. time: ", rank_time_buffer(i,1), &
+            !"s | Time min: ", rank_time_buffer(i,2), "s Time max: " , rank_time_buffer(i,3), "s"
+            write (6, '(a,i3,a,i16,a,f7.2,a,a,f12.2,a)') "MPI Window: ", i, " | Total MC Steps: ", INT(wl_mc_steps(i)), &
+            " | Percent Diff. to Mean: ", 100.0_real64*(wl_mc_steps(i)/(SUM(wl_mc_steps)/SIZE(wl_mc_steps))), "%", &
+            " | Time Taken: ",  rank_time_buffer(i,3), "s"
           end do
           wl_f_prev = wl_f
         end if
@@ -281,6 +289,8 @@ module wang_landau
         if (ANY([0,1] == wl_setup_internal%performance)) then
         call mpi_window_optimise(iter)
         end if
+
+        wl_mc_steps = 0.0_real64
 
         call compute_mean_energy(wl_logdos)
 
@@ -1019,6 +1029,7 @@ module wang_landau
     allocate(diffusion_prev(wl_setup_internal%num_windows))
[…]
