### 2025-11-02 ced74fe
Fixed minor bug that added unnecessary runtime
