### 2025-11-02 0caf723
Changed minimum bins to be 10 for optimiser
