### 2025-11-02 1e39eb0
Fixed legend location
