### 2025-11-03 98c661b
Updated scaling algorithm, removed excess logdos re-zeroing
