### 2025-11-03 3e8f292
Removed weight print
