### 2025-11-03 7b417a6
Made MC step optimise scale with MC steps per bin
