### 2025-11-03 a52ef11
Added walker per instance efficiency plot
