### 2025-11-03 52c7a74
Added horizontal line for efficiency for walkers
