### 2025-11-03 fe1883d
Adjusting sweep plotting script
