### 2025-11-03 7259e37
Adjusting sweep plotting script
