### 2025-11-03 fb87c8b
Adjusting sweep plotting script
