### 2025-11-03 631788c
Adjusting sweep plotting script
