### 2025-11-03 b05e7c5
Adjusting sweep plotting script
