### 2025-11-04 a640388
Adjusted pre_sampling such that it consistently has same minimum regardless of walkers
