### 2025-11-04 6b2d12e
Various adjustments to ensure performance analysis compares like for like
