+++ src/wang-landau.F90
@@ -193,15 +193,14 @@ module wang_landau
     mpi_wl_hist = 0.0_real64
     i = 0
     i_sweeps = 0
-    lb_mc_steps = 0.0_real64
     start = mpi_wtime()
     do while (wl_f > wl_setup_internal%tolerance)
       i_sweeps = i_sweeps + 1
-      if (MOD(i_sweeps, 1000) == 0) then
+      if (MOD(i_sweeps*wl_setup_internal%mc_sweeps, 100000) == 0) then
         i_sweeps = 0
         if (my_rank == 0) then
           print*, ""
-          print*, "1000 Iterations performed. Converged ranks: ", REAL(converged_sum)/REAL(mpi_processes)
+          print*, "100000 Sweeps performed. Converged ranks: ", REAL(converged_sum)/REAL(mpi_processes)
         end if
         call comms_wait()
         if (converged == 0) then
@@ -209,10 +208,11 @@ module wang_landau
           REAL(COUNT(INT(mpi_wl_hist)/=0))/REAL(SIZE(mpi_wl_hist)), "Acceptance Ratio: ", &
           REAL(accepted)/(REAL(wl_setup_internal%mc_sweeps*setup_internal%n_atoms))*100, "%"
         end if
+        mpi_wl_hist = 0.0_real64
       end if 
       call sweeps(wl_logdos, mpi_wl_hist)
 
-      lb_mc_steps(iter+1) = lb_mc_steps(iter+1) + wl_setup_internal%mc_sweeps*setup_internal%n_atoms
+      lb_mc_steps(iter+2) = lb_mc_steps(iter+2) + wl_setup_internal%mc_sweeps*setup_internal%n_atoms
       if (converged == 0) then
         wl_mc_steps(mpi_index) = wl_mc_steps(mpi_index) + wl_setup_internal%mc_sweeps*setup_internal%n_atoms
       end if
@@ -241,7 +241,9 @@ module wang_landau
         ! Increase iteration
         iter = iter + 1
         !Reduce f
-        wl_f = wl_f*1/2
+        wl_f = wl_f*0.5_real64
+        !wl_f = wl_f/(1.0_real64 + 0.5_real64*iter)
+
         !wl_logdos = wl_logdos - minval(wl_logdos, MASK=(wl_logdos > 0.0_real64))!wl_logdos_min
         !wl_logdos = ABS(wl_logdos * merge(0, 1, wl_logdos < 0.0_real64))
         !Average DoS across walkers
@@ -270,12 +272,12 @@ module wang_landau
           end do
           write(*,*)
           write (6, '(a,f20.18,a,f8.2,a)', advance='no') "Flatness reached f of: ", wl_f_prev, &
-                  " | Radial samples: ", radial_min*100_real64, "%"
+                  " | Radial samples: ", radial_min*100.0_real64, "%"
           write (*, *)
           do i=1, wl_setup_internal%num_windows
             !write (6, '(a,i3,a,f12.2,a,f12.2,a,f12.2,a)') "MPI Window: ", i, " | Avg. time: ", rank_time_buffer(i,1), &
             !"s | Time min: ", rank_time_buffer(i,2), "s Time max: " , rank_time_buffer(i,3), "s"
-            write (6, '(a,i3,a,i16,a,f7.2,a,a,f12.2,a)') "MPI Window: ", i, " | Total MC Steps: ", INT(wl_mc_steps(i)), &
+            write (6, '(a,i3,a,f16.0,a,f7.2,a,a,f12.2,a)') "MPI Window: ", i, " | Total MC Steps: ", wl_mc_steps(i), &
[…]
