### 2025-11-06 5ca6403
Adjusted histogram reset after N sweeps
