### 2025-11-07 0e0292e
Removed mid iteration visit histogram adjustment
