### 2025-11-07 #8346 @rkurchin
Okay, apologies again for continued delays here, everyone, and thanks for your patience!

A few specific followups on this:
1. Having performance metrics in a JOSS paper is quite standard, so I have no issue with those being added in.
2. The more detailed theory explainers are something we typically don't have, and this is actually less related to the length limit and more to our policy that we are publishing software tools rather than underlying theories themselves. Obviously, there's no ambiguity in this case that the theory itself is being claimed as new (given it's ~90 years old 🤪), but I still think that stylistically, that theoretical background section doesn't really belong in the JOSS paper and fits better in the docs, which as I read @amkrajewski's [comment](https://github.com/openjournals/joss-reviews/issues/8346#issuecomment-3349814409) above, would still satisfy their concern on that point. @ChrisWoodgate, what would you think about just expanding the docs with some kind of "theory" section and moving that "Bragg-Williams Hamiltonian" section (along with any other details you'd like to from the arXiv posting, though I see the docs already link to that anyway) in there?

Basically, as I envision the outcome here, we would have:
1. performance section from arXiv moved into JOSS paper
2. BW Hamiltonian section moved out and into docs

And at that point, even if the length is still over (which it might be), I'm happy to sign off on all the content being typical and appropriate.
