### 2025-11-10 e421e3b
Reverted load balance to original timing based version with decay
