+++ src/wang-landau.F90
@@ -442,7 +442,7 @@ module wang_landau
     real(real64), allocatable, intent(in) :: rank_time_buffer(:, :), lb_mc_steps(:), window_overlap(:,:)
     call MPI_REDUCE(lb_mc_steps, lb_mc_steps_buffer, SIZE(lb_mc_steps), MPI_DOUBLE_PRECISION, MPI_SUM, 0, MPI_COMM_WORLD, ierr)
     if (my_rank == 0) then
-      lb_bins(iter, :) = REAL(window_indices(:, 2) - window_indices(:, 1) + 1)
+      lb_bins(iter, :) = REAL(window_intervals(:, 2) - window_intervals(:, 1) + 1)
       lb_avg_time(iter, :) = rank_time_buffer(:,1)
       lb_max_time(iter, :) = rank_time_buffer(:,3)
       window_time(iter) = MAXVAL(rank_time_buffer(:,3))
@@ -1112,7 +1112,7 @@ module wang_landau
     target_energy = (mpi_bin_edges(1) + mpi_bin_edges(SIZE(mpi_bin_edges)))/2
 
     ! Load balance diffusion
-    diffusion_prev = (window_intervals(:,2) - window_intervals(:,1) + 1)**2
+    diffusion_prev = 1.0_real64/REAL(wl_setup_internal%num_windows)
 
     ! Load balance arrays
     lb_bins = -1.0_real64
@@ -1162,37 +1162,10 @@ module wang_landau
   !> @date    2024 
   subroutine dos_average(wl_logdos)
     real(real64), intent(inout) :: wl_logdos(:)
-    integer :: i, j
 
-    do i = 1, wl_setup_internal%num_windows
-      if (mpi_index == i) then
-        if (my_rank /= (i - 1)*num_walkers) then
-          call MPI_Send(wl_logdos, wl_setup_internal%bins, MPI_DOUBLE_PRECISION, (i - 1)*num_walkers, i, MPI_COMM_WORLD, ierr)
-          !print*, my_rank, "send", (i - 1)*num_walkers
-        end if
-        if (my_rank == (i - 1)*num_walkers) then
-          do j = 1, num_walkers - 1
-            call MPI_Recv(wl_logdos_buffer, wl_setup_internal%bins, MPI_DOUBLE_PRECISION, &
-            (i - 1)*num_walkers + j, i, MPI_COMM_WORLD, MPI_STATUS_IGNORE, ierr)
-            !print*, my_rank, "recv", (i - 1)*num_walkers + j
-            wl_logdos = wl_logdos + wl_logdos_buffer
-          end do
-        end if
-        if (my_rank == (i - 1)*num_walkers) then
-          wl_logdos = wl_logdos/num_walkers
-          do j = 1, num_walkers - 1
-            call MPI_Send(wl_logdos, wl_setup_internal%bins, MPI_DOUBLE_PRECISION, (i - 1)*num_walkers + j, i, MPI_COMM_WORLD, &
-                          ierr)
-            !print*, my_rank, "send", (i - 1)*num_walkers + j
-          end do
-        else
-          call MPI_Recv(wl_logdos_buffer, wl_setup_internal%bins, MPI_DOUBLE_PRECISION, (i - 1)*num_walkers, i, MPI_COMM_WORLD, &
-                        MPI_STATUS_IGNORE, ierr)
-          wl_logdos = wl_logdos_buffer
-          !print*, my_rank, "recv", (i - 1)*num_walkers
-        end if
-      end if
-    end do
+    call MPI_Allreduce(MPI_IN_PLACE, wl_logdos, size(wl_logdos), MPI_DOUBLE_PRECISION, MPI_SUM, sub_comm, ierr)
[…]
