### 2025-11-11 a1e04aa
Removed non functioning plotting script
