+++ examples/plotting_scripts/performance_plotting.py
@@ -444,159 +444,3 @@ for overlap in range(len(overlaps)):
 # Walker Efficiency
 
 plot_walker_efficiency([0, 1, 2], [0], replica=True)
-
-# Sweep comparison
-methods = np.array([0, 2, 4])
-methods_label = np.array(["1", "3", "5"])
-windows = np.array([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16])
-walkers = np.array([1, 6])
-overlaps = np.array([25])
-sweeps = np.array([1, 50, 200])
-
-for sweep in sweeps:
-    sums = np.zeros([len(methods), len(walkers), len(windows), len(overlaps), repeats])
-    for method_id, method in enumerate(methods):
-        for walker in range(len(walkers)):
-            for window in range(len(windows)):
-                for overlap in range(len(overlaps)):
-                    for k in range(1, repeats + 1):
-                        if walkers[walker] == 6:
-                            filename = (
-                                "sharing_{}_{:02d}_{:02d}_{:02d}_{:03d}_{}/load_balance/wl_lb_max_time.dat"
-                                .format(method, walkers[walker], windows[window],
-                                        overlaps[overlap], sweep, k)
-                            )
-                            wl_lb_max_time = nc.Dataset(filename)
-                            wl_lb_max_time = np.array(
-                                wl_lb_max_time["grid data"][:], dtype=np.float64
-                            ).T
-                        elif walkers[walker] == 6:
-                            filename = (
-                                "{}_{:02d}_{:02d}_{:02d}_{}/load_balance/wl_lb_max_time.dat"
-                                .format(method, walkers[walker], windows[window],
-                                        overlaps[overlap], k)
-                            )
-                            wl_lb_max_time = nc.Dataset(filename)
-                            wl_lb_max_time = np.array(
-                                wl_lb_max_time["grid data"][:], dtype=np.float64
-                            ).T
-
-                        row_sums = np.max(wl_lb_max_time, axis=1)
-                        sums[method_id, walker, window, overlap, k - 1] = np.sum(row_sums)
-
-    sums_mean = np.mean(sums, axis=4)
-    sums_min = np.min(sums, axis=4)
-    sums_err = np.std(sums, axis=4) / np.sqrt(repeats)
-
-    one_sums = np.zeros([len(methods), len(windows), len(overlaps), repeats])
-    one_sums = sums[:, 0, :, :, :]
-    one_sums_err = np.zeros([len(windows), len(overlaps)])
-    one_sums_mean = np.zeros([len(windows), len(overlaps)])
-
-    for window in range(len(windows)):
-        one_sums_err[window] = np.std(one_sums[:, window, :]) / np.sqrt(len(methods) * repeats)
-        one_sums_mean[window] = np.mean(one_sums[:, window, :])
-
-        sums_mean[:, 0, window] = one_sums_mean[window]
-        sums_err[:, 0, window] = one_sums_err[window]
-
-    for window in range(len(windows)):
-        for overlap in range(len(overlaps)):
-            one_sums_err[window, overlap] = (
-                np.std(one_sums[:, window, overlap, :]) /
-                np.sqrt(len(methods) * repeats)
-            )
-            one_sums_mean[window, overlap] = np.mean(one_sums[:, window, overlap, :])
-
-            sums_mean[:, 0, window, overlap] = one_sums_mean[window, overlap]
-            sums_err[:, 0, window, overlap] = one_sums_err[window, overlap]
-
-    efficiency = np.empty_like(sums_mean)
-    efficiency_err = np.empty_like(sums_err)
-
-    for method in range(len(methods)):
-        for overlap in range(len(overlaps)):
-            for walker in range(len(walkers)):
[…]
