### 2025-11-13 #8346 @ChrisWoodgate
Hi @rkurchin,

No problem at all regarding the delays—thanks for taking the time to investigate this. 🙂

Frankly, at this point, I'm happy to go with whatever I have to do to get this off my desk! 😅 So, I think your proposed solution of

> 1. performance section from arXiv moved into JOSS paper
> 2. BW Hamiltonian section moved out and into docs

works fine for me. Personally, I don't like putting theoretical background in code documentation, but if that's what we have to do in this case, so be it.

I will talk to @HubertJN and work on this at some point over the next week or two. I am slammed with deadlines at the moment so it may take a while, particularly as "BW Hamiltonian section moved out and into docs" will require us thoroughly checking that the article stands on its own once removed.
