### 2025-11-19 83aecd9
Cutting out background theory to comply with editorial guidance
