+++ paper/paper.md
@@ -44,7 +44,7 @@ Frequently, a mixture of elements will form a regular crystalline lattice with s
 The nature of atomic arrangements in a material determines many important physical properties.
 For a given combination of elements, it is therefore crucial to understand the nature of atomic ordering in a material, as well as the temperature at which it emerges upon cooling, to guide materials processing strategies.
 One physically intuitive model for the internal energy of an alloy is the Bragg-Williams model, which assumes that atoms in the alloy interact in a pairwise manner.
-Crucially, the effective pair interactions (EPIs) which appear in the Bragg-Williams Hamiltonian can be obtained _ab initio_ using density functional theory (DFT) calculations.
+Crucially, the atom-atom effective pair interactions (EPIs) which appear in the Bragg-Williams Hamiltonian can be obtained _ab initio_ using density functional theory (DFT) calculations.
 When appropriate sampling techniques are applied to the Bragg-Williams model, it is possible to explore the configuration space of a given alloy in detail and determine equilibrium phases as a function of temperature, leading to construction of phase diagrams.
 Here, we present `BraWl`, a Fortran package implementing a range of conventional and enhanced sampling algorithms for exploration of the phase space of the Bragg-Williams model, facilitating study of diffusional solid-solid transformations in binary and multicomponent alloys.
 These sampling algorithms include Metropolis-Hastings Monte Carlo, Wang-Landau sampling, and Nested Sampling.
@@ -63,35 +63,6 @@ Examples of widely-used such packages include ATAT [@vandewallealloy2002], ICET 
 To our knowledge, there is no open-source package specifically focussing on the implementation of a range of sampling algorithms applied to the Bragg-Williams model.
 We therefore believe that BraWl fills a gap in the capabilities of the current alloy software ecosystem.
 Additionally, we hope that the modular way in which the package is constructed could enable implementation of more complex Hamiltonians, as well as further sampling algorithms in addition to those detailed below, in due course.
-
-## The Bragg-Williams Hamiltonian
-
-A physically intuitive, lattice-based model for the internal energy of a substitutional alloy is the Bragg-Williams model [@braggeffect1934; @braggeffect1935], which assumes that the internal energy of an alloy takes a simple, pairwise form.
-We specify a particular arrangement of atoms by a discrete set of site occupation numbers, $\left\{ \xi_{i\gamma} \right\}$, where $\xi_{i \gamma} = 1$ if lattice site $i$ is occupied by an atom of species $\gamma$ and $\xi_{i \gamma} = 0$ otherwise.
-The Bragg-Williams Hamiltonian then has the form
-\begin{equation}
-    H(\{\xi_{i\gamma}\}) = \frac{1}{2}\sum_{i \gamma; j\gamma^{\prime}} V_{i\gamma; j\gamma^{\prime}} \xi_{i \gamma} \xi_{j \gamma^{\prime}},
-    \label{eq:b-w1}
-\end{equation}
-where $V_{i\gamma; j\gamma^{\prime}}$ denotes the effective pair interaction (EPI) between an atom of chemical species $\gamma$ on lattice site $i$ and an atom of chemical species $\gamma^{\prime}$ on lattice site $j$.
-(The factor of $\frac{1}{2}$ eliminates double-counting in the summation.)
-For a system of finite size, it is assumed that periodic boundary conditions are applied in all coordinate directions.
-
-Generally, the assumption is made that that the EPIs are spatially homogeneous and isotropic, and \autoref{eq:b-w1} is therefore rewritten as
-\begin{equation}
[…]
