+++ paper/brawl.bib
@@ -30,6 +30,20 @@
 	pages = {115124},
 }
 
+@book{woodgatemodelling2024,
+	address = {Cham},
+	series = {Springer {Series} in {Materials} {Science}},
+	title = {Modelling {Atomic} {Arrangements} in {Multicomponent} {Alloys}: {A} {Perturbative}, {First}-{Principles}-{Based} {Approach}},
+	volume = {346},
+	isbn = {978-3-031-62020-1 978-3-031-62021-8},
+	shorttitle = {Modelling {Atomic} {Arrangements} in {Multicomponent} {Alloys}},
+	url = {https://doi.org/10.1007/978-3-031-62021-8},
+	urldate = {2024-08-12},
+	publisher = {Springer Nature Switzerland},
+	author = {Woodgate, Christopher D.},
+	year = {2024},
+}
+
 @article{woodgatecompetition2024,
 	title = {Competition between phase ordering and phase segregation in the {Ti}$_{\textrm{x}}${NbMoTaW} and {Ti}$_{\textrm{x}}${VNbMoTaW} refractory high-entropy alloys},
 	volume = {135},
@@ -57,45 +71,6 @@
 	month = mar,
 	year = {2024},
 	pages = {033804},
-}
-
-@article{rubanatomic2004,
-	title = {Atomic and magnetic configurational energetics by the generalized perturbation method},
-	volume = {70},
-	issn = {1098-0121, 1550-235X},
-	url = {https://doi.org/10.1103/PhysRevB.70.125115},
-	doi = {10.1103/PhysRevB.70.125115},
-	number = {12},
-	urldate = {2023-01-11},
-	journal = {Physical Review B},
-	author = {Ruban, A. V. and Shallcross, S. and Simak, S. I. and Skriver, H. L.},
-	month = sep,
-	year = {2004},
-	pages = {125115},
-}
-
-@article{braggeffect1934,
-	title = {The effect of thermal agitation on atomic arrangement in alloys},
-	volume = {145},
-	url = {https://doi.org/10.1098/rspa.1934.0132},
-	doi = {10.1098/rspa.1934.0132},
-	number = {855},
-	journal = {Proceedings of the Royal Society of London. Series A, Containing Papers of a Mathematical and Physical Character},
-	author = {Bragg, W. L. and Williams, E. J.},
-	year = {1934},
-	pages = {699--730},
-}
-
-@article{braggeffect1935,
-	title = {The effect of thermal agitaion on atomic arrangement in alloys-{II}},
-	volume = {151},
-	url = {https://doi.org/10.1098/rspa.1935.0165},
-	doi = {10.1098/rspa.1935.0165},
-	number = {874},
-	journal = {Proceedings of the Royal Society of London. Series A - Mathematical and Physical Sciences},
-	author = {Bragg, W. L. and Williams, E. J.},
-	year = {1935},
-	pages = {540--566},
 }
 
 @article{woodgateshortrange2023,
@@ -347,36 +322,6 @@
 	pages = {045002},
 }
 
-@article{ducastellegeneralized1976,
-	title = {Generalized perturbation theory in disordered transitional alloys: {Applications} to the calculation of ordering energies},
-	volume = {6},
-	issn = {0305-4608},
-	shorttitle = {Generalized perturbation theory in disordered transitional alloys},
-	url = {https://doi.org/10.1088/0305-4608/6/11/005},
-	doi = {10.1088/0305-4608/6/11/005},
-	number = {11},
-	urldate = {2025-03-25},
-	journal = {Journal of Physics F: Metal Physics},
-	author = {Ducastelle, F and Gautier, F},
-	month = nov,
-	year = {1976},
-	pages = {2039--2062},
-}
-
[…]
