### 2025-11-19 a2e4942
Adding in 'Performance' section to comply with editorial guidance.
