+++ paper/paper.md
@@ -61,7 +61,7 @@ An in-depth discussion of the background, underlying theory, and technical detai
 There are a range of existing packages capable of simulating alloy phase equilibria, both open- and closed-source.
 Examples of widely-used such packages include ATAT [@vandewallealloy2002], ICET [@angqvisticet2019] and CELL [@rigamonticell2024], though all of these focus primarily on implementation of a generalised cluster expansion, rather than the simpler form of the Bragg-Williams Hamiltonian.
 To our knowledge, there is no open-source package specifically focussing on the implementation of a range of sampling algorithms applied to the Bragg-Williams model.
-We therefore believe that BraWl fills a gap in the capabilities of the current alloy software ecosystem.
+We therefore believe that `BraWl` fills a gap in the capabilities of the current alloy software ecosystem.
 Additionally, we hope that the modular way in which the package is constructed could enable implementation of more complex Hamiltonians, as well as further sampling algorithms in addition to those detailed below, in due course.
 
 # Sampling algorithms
@@ -80,7 +80,7 @@ At present, the implemented sampling algorithms are:
 The relevant quantities are: 
 
 - **Internal energy.** For a given lattice type, system size, alloy composition, and set of atom-atom effective pair interactions, `BraWl` can evaluate the total energy associated with the alloy configuration (\autoref{eq:b-w1}). At the time of writing, for speed, common lattice types (fcc, bcc, simple cubic, _etc._) are hard-coded, with the intention that the range of implemented lattice types will be expanded over time as necessary.
-- **Heat capacity.** The isochoric (fixed volume) heat capacity at a given temperature, $C_V(T)$, is a useful quantity for identifying phase transitions, as a plot of the simulation heat capacity as a function of temperature is expected to show a local peak at the temperature at which the transition occurs. Within `BraWl`, the heat capacity is calculated via $$C_V(T) = \frac{\langle E^2\rangle - \langle E\rangle^2}{k_BT^2},$$ where $k_B$ is the usual Boltzmann constant, $E$ is the simulation energy, and $\langle \cdot \rangle$ denote thermodynamic averages obtained using the relevant sampling algorithm.
+- **Heat capacity.** The isochoric (fixed volume) specific heat at a given temperature, $C_V(T)$, is a useful quantity for identifying phase transitions, as a plot of the specific heat as a function of temperature is expected to show a local peak at the temperature at which the transition occurs. Within `BraWl`, the specific heat is calculated via $$C_V(T) = \frac{\langle E^2\rangle - \langle E\rangle^2}{k_BT^2},$$ where $k_B$ is the usual Boltzmann constant, $E$ is the simulation energy, and $\langle \cdot \rangle$ denote thermodynamic averages obtained using the relevant sampling algorithm.
 - **Atomic short-range order parameters.** To assess local atom-atom correlations in a simulation, `BraWl` can calculate the Warren-Cowley atomic short-range order parameters [@cowleyapproximate1950; @cowleyshortrange1965], adapted to the multicomponent setting, defined as $$\alpha^{\gamma \gamma'}_n=1-\frac{P^{\gamma \gamma'}_n}{c_{\gamma'}},$$ where $n$ refers to the $n$th coordination shell, $P^{\gamma \gamma'}_n$ is the conditional probability of an atom of type $q$ neighbouring an atom of type $p$ on shell $n$, and $c_q$ is the total concentration of atom type $q$. When $\alpha^{\gamma \gamma'}_n > 0$, $p$-$q$ pairs are disfavoured on shell $n$ and, when $\alpha^{\gamma \gamma'}_n < 0$ they are favoured. The value $\alpha^{\gamma \gamma'}_n = 0$ corresponds to the ideal, maximally disordered solid solution.
[…]
