+++ paper/brawl.bib
@@ -38,6 +38,7 @@
 	isbn = {978-3-031-62020-1 978-3-031-62021-8},
 	shorttitle = {Modelling {Atomic} {Arrangements} in {Multicomponent} {Alloys}},
 	url = {https://doi.org/10.1007/978-3-031-62021-8},
+	doi = {10.1007/978-3-031-62021-8},
 	urldate = {2024-08-12},
 	publisher = {Springer Nature Switzerland},
 	author = {Woodgate, Christopher D.},
@@ -376,7 +377,7 @@
 	doi = {10.1038/s41524-024-01363-x},
 	number = {1},
 	urldate = {2025-04-09},
-	journal = {npj Computational Materials},
+	journal = {{npj} Computational Materials},
 	author = {Rigamonti, Santiago and Troppenz, Maria and Kuban, Martin and Hübner, Axel and Draxl, Claudia},
 	month = aug,
 	year = {2024},
@@ -406,7 +407,7 @@
 }
 
 @article{fisherlattice2025,
-	title = {Lattice vacancy migration barriers in  Fe-Ni alloys, and why Ni atoms diffuse slowly: {An} \textit{ab initio} study},
+	title = {Lattice vacancy migration barriers in {Fe-Ni} alloys, and why {Ni} atoms diffuse slowly: {An} \textit{ab initio} study},
 	copyright = {arXiv.org perpetual, non-exclusive license},
 	url = {https://doi.org/10.48550/arXiv.2508.19124},
 	urldate = {2025-08-27},
