### 2025-11-19 4ec7217
Enlarging figures
