+++ paper/brawl.bib
@@ -125,8 +125,9 @@
 	address = {Cambridge, UK},
 	edition = {4},
 	title = {A {Guide} to {Monte} {Carlo} {Simulations} in {Statistical} {Physics}},
+	isbn = {978-1-139-69646-3},
 	url = {https://doi.org/10.1017/CBO9781139696463},
-    doi = {10.1017/CBO9781139696463},
+        doi = {10.1017/CBO9781139696463},
 	publisher = {Cambridge University Press},
 	author = {Landau, David P. and Binder, Kurt},
 	year = {2014},
