+++ paper/paper.md
@@ -71,7 +71,7 @@ Additionally, we hope that the modular way in which the package is constructed c
 Throughout, the package defaults to performing _swaps_ of pairs of atoms in the simulation cell, to conserve the overall concentration of each chemical species present in the simulation.
 At present, the implemented sampling algorithms are:
 
-1. **The Metropolis-Hastings Monte Carlo algorithm.** The Metropolis-Hastings algorithm allows for a system of interest to follow a chain of states which evolve to, and sample, an equilibrium ensemble [@metropolisequation1953; landauguide2014]. It is a useful method for obtaining the equilibrium state of a system at a given temperature.
+1. **The Metropolis-Hastings Monte Carlo algorithm.** The Metropolis-Hastings algorithm allows for a system of interest to follow a chain of states which evolve to, and sample, an equilibrium ensemble [@metropolisequation1953; @landauguide2014]. It is a useful method for obtaining the equilibrium state of a system at a given temperature.
 2. **Wang-Landau sampling.** Wang-Landau sampling is a flat histogram method which provides a means for high throughput calculation of phase diagrams for atomistic/lattice model systems [@wangefficient2001]. The method allows for direct computation of an estimate of the density of states in energy, $g(E)$, and hence the partition function of a system of interest. From the partition function, thermodynamic quantities at any temperature of interest can then be obtained provided one has prior knowledge of the minimum and maximum energy relevant to those temperatures.
 3. **Nested Sampling.** Nested sampling (NS) is powerful Bayesian inference technique [@skillingnested2004; @skillingnested2006] adapted to sample the potential energy surface of atomistic systems [@ashtonnested2022], giving direct access to the partition function at arbitrary temperatures for comprehensive thermodynamic analysis, without relying on advance knowledge of relevant structures or the range of energies accessible to them [@partayefficient2010; @partaynested2021].
 
