### 2025-11-21 9255e18
Tweaks suggested by David
