+++ paper/paper.md
@@ -90,16 +90,15 @@ The relevant quantities are:
 BraWL has been used, with success, to study the phase behaviour of a range of binary and multicomponent alloys, for example the binary Fe-Ga system (Galfenol) [@marchantab2021], the Fe-Ni system [@woodgateint2024], the Cantor-Wu medium- and high-entropy alloys [@woodgatecompositional2022; @woodgateinterplay2023], the refractory high-entropy alloys [@woodgateshortrange2023; @woodgatecompetition2024], the Al$_x$CrFeCoNi system [@woodgatestructure2024], and the AlTiVNb and AlTiCrMo refractory high-entropy superalloys [@woodgateemergent2025].
 Additionally, the package has been used to generate atomic configurations with physically motivated ASRO and/or ALRO for subsequent study using a range of other simulation techniques.
 We highlight examples of its use in generating a training dataset for a machine-learned interatomic potential for the prototypical austenitic stainless steel, Fe$_7$Cr$_2$Ni [@shenoycollinearspin2024], as well as its use in generating configurations for use in a transition state study for ferromagnetic Fe-Ni alloys [@fisherlattice2025].
-Finally, the package has also been used to benchmark the efficiency of various parallelisation strategies proposed for the Wang-Landau sampling algorithm [@naguszewskioptimal2025].
 In this work, we explicitly consider several illustrative examples of the results which can be obtained using the sampling algorithms outlined above applied to the Bragg-Williams model as implemented in the package. Throughout these examples, the atom-atom EPIs for the Bragg-Williams model are obtained using the $S^{(2)}$ theory for multicomponent alloys [@khanstatistical2016; @woodgatemodelling2024].
 
 As an example of the Metropolis-Hastings Monte Carlo algorithm, we consider its application to the binary FeNi alloy, first discussed by @woodgateint2024.
 \autoref{fig:feniequilibration} shows the internal energy and conditional pair probabilities (quantifying ASRO) of a simulation cell containing 256 atoms as a function of the number of Metropolis-Hastings 'sweeps', where a sweep refers to performing a number of trial Metropolis-Hastings moves equal to the number of atoms in the simulation cell.
-The simulation is performed at 300 K, below the alloy's L1$_0$ disorder-order transition temperature.
-The L1$_0$ phase is a structure where $2/3$ of the nearest neighbours of Fe (Ni) atoms are Ni (Fe) atoms, and where none of the next-nearest neighbours of Fe (Ni) atoms are Ni (Fe) atoms.
+The simulation is performed at 300 K, below the alloy's $\textrm{L}1_0$ disorder-order transition temperature.
+The $\textrm{L}1_0$ phase is a structure where $2/3$ of the nearest neighbours of Fe (Ni) atoms are Ni (Fe) atoms, and where none of the next-nearest neighbours of Fe (Ni) atoms are Ni (Fe) atoms.
 It can be seen that this ordering is swiftly established as the number of Monte Carlo sweeps increases, albeit with some remaining thermal noise.
 
-![Evolution of the simulation internal energy (top panel) and conditional pair probabilities (bottom panel) for an Fe$_{0.5}$Ni$_{0.5}$ alloy as a function of the number of Metropolis-Hastings sweeps at a simulation temperature of $T=300$ K. One 'sweep' is one trial move per atom in the system. Beyond approximately 100 sweeps, the system can be seen to have reached equilibrium, with L1$_0$ order established.\label{fig:feniequilibration}](feniequilibration.pdf){ width=70% }
[…]
