+++ paper/brawl.bib
@@ -417,3 +417,18 @@
         doi = {10.48550/arXiv.2508.19124},
 	author = {Fisher, Adam M. and Woodgate, Christopher D. and Xiaoyu Zhang and George C. Hadjipanayis and Lewis, Laura H. and Staunton, Julie B.},
 }
+
+@article{unglertreplica2025,
+	title = {Replica {Exchange} {Nested} {Sampling}},
+	volume = {21},
+	issn = {1549-9618, 1549-9626},
+	url = {https://doi.org/10.1021/acs.jctc.5c00588},
+	doi = {10.1021/acs.jctc.5c00588},
+	number = {15},
+	urldate = {2025-11-21},
+	journal = {Journal of Chemical Theory and Computation},
+	author = {Unglert, N. and Pártay, L. B. and Madsen, G. K. H.},
+	month = aug,
+	year = {2025},
+	pages = {7304--7319},
+}
