### 2025-11-21 1078a54
Fixing consistency of 'Specific Heat' and 'Heat Capacity'
