+++ paper/paper.md
@@ -81,7 +81,7 @@ At present, the implemented sampling algorithms are:
 The relevant quantities are: 
 
 - **Internal energy.** For a given lattice type, system size, alloy composition, and set of atom-atom effective pair interactions, `BraWl` can evaluate the total energy associated with the alloy configuration. At the time of writing, for speed, common lattice types (fcc, bcc, simple cubic, _etc._) are hard-coded, with the intention that the range of implemented lattice types will be expanded over time as necessary.
-- **Heat capacity.** The isochoric (fixed volume) specific heat at a given temperature, $C_V(T)$, is a useful quantity for identifying phase transitions, as a plot of the specific heat as a function of temperature is expected to show a local peak at the temperature at which the transition occurs. Within `BraWl`, the specific heat is calculated via $$C_V(T) = \frac{\langle E^2\rangle - \langle E\rangle^2}{k_BT^2},$$ where $k_B$ is the usual Boltzmann constant, $E$ is the simulation energy, and $\langle \cdot \rangle$ denote thermodynamic averages obtained using the relevant sampling algorithm.
+- **Specific heat.** The isochoric (fixed volume) specific heat at a given temperature, $C_V(T)$, is a useful quantity for identifying phase transitions, as a plot of the specific heat as a function of temperature is expected to show a local peak at the temperature at which the transition occurs. Within `BraWl`, the specific heat is calculated via $$C_V(T) = \frac{\langle E^2\rangle - \langle E\rangle^2}{k_BT^2},$$ where $k_B$ is the usual Boltzmann constant, $E$ is the simulation energy, and $\langle \cdot \rangle$ denote thermodynamic averages obtained using the relevant sampling algorithm.
 - **Atomic short-range order parameters.** To assess local atom-atom correlations in a simulation, `BraWl` can calculate the Warren-Cowley atomic short-range order parameters [@cowleyapproximate1950; @cowleyshortrange1965], adapted to the multicomponent setting, defined as $$\alpha^{\gamma \gamma'}_n=1-\frac{P^{\gamma \gamma'}_n}{c_{\gamma'}},$$ where $n$ refers to the $n$th coordination shell, $P^{\gamma \gamma'}_n$ is the conditional probability of an atom of type $q$ neighbouring an atom of type $p$ on shell $n$, and $c_q$ is the total concentration of atom type $q$. When $\alpha^{\gamma \gamma'}_n > 0$, $p$-$q$ pairs are disfavoured on shell $n$ and, when $\alpha^{\gamma \gamma'}_n < 0$ they are favoured. The value $\alpha^{\gamma \gamma'}_n = 0$ corresponds to the ideal, maximally disordered solid solution.
 - **Atomic long-range order parameters.** Over a simulation run (for example using the Metropolis algorithm), `BraWl` can calculate the average occupancy of a lattice site, _i.e._ the probability of an atom of a particular chemical species sitting on that site. This capability was first demonstrated on simulations of Fe-Ga alloys [@marchantab2021].
 
