### 2025-11-25 2bc03c1
Minor wording tweak.
