### 2025-11-25 e36007b
Fixing minor typo in README.md
