### 2025-11-25 #8346 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
