+++ paper/paper.md
@@ -22,7 +22,7 @@ authors:
     orcid: 0000-0003-4876-5558
     affiliation: 3
 affiliations:
- - name: Department of Physics, University of Warwick, UK
+ - name: Department of Physics, University of Warwick, United Kingdom
    index: 1
    ror: 01a77tt86
  - name: Department of Chemistry, University of Warwick, UK
@@ -53,7 +53,7 @@ We demonstrate the capabilities of the package by applying it to some prototypic
 # Statement of need
 
 The Fortran package `BraWl` facilitates simulation of the thermodynamics and phase stability of both binary and multicomponent alloys.
-It achieves this by providing implementation of both the Bragg-Williams Hamiltonian (a lattice based model expressing the internal energy of an alloy as a sum of atom-atom effective pair interactions) concurrently with a range of conventional and enhanced sampling techniques for exploration of the alloy configuration space.
+It achieves this by providing implementation of both the Bragg-Williams Hamiltonian (a lattice-based model expressing the internal energy of an alloy as a sum of atom-atom effective pair interactions) concurrently with a range of conventional and enhanced sampling techniques for exploration of the alloy configuration space.
 The result is a package which can determine phase equilibria as a function of both temperature and alloy composition, which leads to the construction of alloy phase diagrams.
 Additionally, the package can be used for extraction of representative equilibrated atomic configurations for visualisation, as well as for use in complementary modelling approaches.
 An in-depth discussion of the background, underlying theory, and technical details of the sampling algorithms implemented in the package is given by @naguszewskibrawl2025.
@@ -72,7 +72,7 @@ Throughout, the package defaults to performing _swaps_ of pairs of atoms in the 
 At present, the implemented sampling algorithms are:
 
 1. **The Metropolis-Hastings Monte Carlo algorithm.** The Metropolis-Hastings algorithm allows for a system of interest to follow a chain of states which evolve to, and sample, an equilibrium ensemble [@metropolisequation1953; @landauguide2014]. It is a useful method for obtaining the equilibrium state of a system at a given temperature.
-2. **Wang-Landau sampling.** Wang-Landau sampling is a flat histogram method which provides a means for high throughput calculation of phase diagrams for atomistic/lattice model systems [@wangefficient2001]. The method allows for direct computation of an estimate of the density of states in energy, $g(E)$, and hence the partition function of a system of interest. From the partition function, thermodynamic quantities at any temperature of interest can then be obtained provided one has prior knowledge of the minimum and maximum energy relevant to those temperatures.
+2. **Wang-Landau sampling.** Wang-Landau sampling is a flat histogram method which provides a means for high-throughput calculation of phase diagrams for atomistic/lattice model systems [@wangefficient2001]. The method allows for direct computation of an estimate of the density of states in energy, $g(E)$, and hence the partition function of a system of interest. From the partition function, thermodynamic quantities at any temperature of interest can then be obtained provided one has prior knowledge of the minimum and maximum energy relevant to those temperatures.
[…]
