### 2025-11-26 8ba0ef6
Additional JOSS tweak to paper.md
